+++ src/components/EditStopPage/ModalitiesMenuItems.js
@@ -17,8 +17,8 @@ import MenuItem from "@mui/material/MenuItem";
 import ModalityIconSvg from "../MainPage/ModalityIconSvg";
 import {
   getInverseSubmodesWhitelist,
-  getStopPlacesForSubmodes,
-} from "../../models/submodes";
+  getStopTypesForSubmodes,
+} from "../../modelUtils/modeUtils";
 import MoreMenuItem from "../MainPage/MoreMenuItem";
 
 class ModalitiesMenuItems extends React.Component {
@@ -38,7 +38,7 @@ class ModalitiesMenuItems extends React.Component {
     const blacklistedStopTypes = allowsInfo.blacklistedStopPlaceTypes || [];
     const illegalSubmodes = getInverseSubmodesWhitelist(legalSubmodes);
     // stopPlacesTypes that submodes are depending on to be legal in order to render
-    const adHocStopPlaceTypes = getStopPlacesForSubmodes(legalSubmodes);
+    const adHocStopPlaceTypes = getStopTypesForSubmodes(legalSubmodes);
     const chosenStyle = { fontWeight: "bold" };
     const unchosenStyle = {};
