+++ src/test/mock/equippedStopPlace.js
