--- src/test/mock/polygon-vestfold.js
@@ -1,3302 +1,0 @@
-/*
- *  Licensed under the EUPL, Version 1.2 or – as soon they will be approved by
-the European Commission - subsequent versions of the EUPL (the "Licence");
-You may not use this work except in compliance with the Licence.
-You may obtain a copy of the Licence at:
-
-  https://joinup.ec.europa.eu/software/page/eupl
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the Licence is distributed on an "AS IS" basis,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the Licence for the specific language governing permissions and
-limitations under the Licence. */
-
-export default [
-  [9.8468, 59.2649],
-  [9.8456, 59.2654],
-  [9.8443, 59.2663],
-  [9.8427, 59.2674],
-  [9.8397, 59.269],
-  [9.8382, 59.2696],
-  [9.8362, 59.272],
-  [9.8299, 59.2745],
-  [9.826, 59.2761],
-  [9.823, 59.2772],
-  [9.8197, 59.2786],
-  [9.8191, 59.2788],
-  [9.8189, 59.2789],
-  [9.8182, 59.2792],
-  [9.817, 59.2797],
-  [9.8132, 59.2812],
-  [9.8106, 59.2823],
-  [9.8106, 59.2823],
-  [9.8106, 59.2861],
-  [9.8106, 59.2877],
-  [9.8106, 59.289],
-  [9.8081, 59.2911],
-  [9.8047, 59.2939],
-  [9.8037, 59.2948],
-  [9.8036, 59.2948],
-  [9.8036, 59.2949],
-  [9.8035, 59.2949],
-  [9.8035, 59.2949],
-  [9.8016, 59.2964],
-  [9.7863, 59.3086],
-  [9.7802, 59.3134],
-  [9.7768, 59.3161],
-  [9.7606, 59.3227],
-  [9.7579, 59.3238],
-  [9.7553, 59.3249],
-  [9.7553, 59.3249],
-  [9.7554, 59.3249],
-  [9.7555, 59.3251],
-  [9.7557, 59.3253],
-  [9.757, 59.3267],
-  [9.7584, 59.3282],
-  [9.7618, 59.332],
-  [9.763, 59.3333],
-  [9.7715, 59.3427],
-  [9.7719, 59.343],
-  [9.7731, 59.3438],
-  [9.7736, 59.3441],
-  [9.7828, 59.3503],
-  [9.7782, 59.3562],
-  [9.7769, 59.3577],
-  [9.7756, 59.3594],
-  [9.7732, 59.363],
-  [9.7719, 59.3673],
-  [9.7706, 59.3716],
-  [9.7698, 59.3745],
-  [9.7698, 59.3745],
-  [9.7698, 59.3745],
-  [9.7697, 59.3745],
-  [9.7694, 59.3755],
-  [9.7694, 59.3756],
-  [9.7694, 59.3756],
-  [9.7698, 59.3764],
-  [9.7707, 59.3786],
-  [9.7707, 59.3786],
-  [9.7707, 59.3786],
-  [9.7767, 59.3807],
-  [9.7767, 59.3807],
-  [9.7771, 59.3824],
-  [9.7774, 59.3833],
-  [9.7779, 59.385],
-  [9.7784, 59.387],
-  [9.7791, 59.3892],
-  [9.7793, 59.39],
-  [9.7794, 59.3904],
-  [9.7794, 59.3905],
-  [9.7793, 59.391],
-  [9.7793, 59.3919],
-  [9.7792, 59.393],
-  [9.7788, 59.3993],
-  [9.7786, 59.4027],
-  [9.7782, 59.4103],
-  [9.7781, 59.412],
-  [9.778, 59.4124],
-  [9.7779, 59.4126],
-  [9.7778, 59.4129],
-  [9.7777, 59.4132],
-  [9.7776, 59.4135],
-  [9.7775, 59.4138],
-  [9.7774, 59.4141],
-  [9.7774, 59.4144],
-  [9.7773, 59.4146],
-  [9.7772, 59.4149],
-  [9.7771, 59.4152],
-  [9.777, 59.4155],
-  [9.7769, 59.4158],
-  [9.7768, 59.4161],
-  [9.7768, 59.4164],
-  [9.7767, 59.4166],
-  [9.7766, 59.417],
-  [9.7765, 59.4173],
-  [9.7764, 59.4176],
-  [9.7763, 59.4178],
-  [9.7762, 59.4181],
-  [9.7761, 59.4184],
-  [9.7761, 59.4187],
-  [9.776, 59.4189],
-  [9.7759, 59.4192],
-  [9.7759, 59.4192],
-  [9.7758, 59.4195],
-  [9.7758, 59.4196],
-  [9.7756, 59.4202],
-  [9.7755, 59.4205],
-  [9.7754, 59.4207],
-  [9.7754, 59.4209],
-  [9.7753, 59.4212],
-  [9.7752, 59.4214],
-  [9.7752, 59.4217],
-  [9.7751, 59.4219],
-  [9.775, 59.4221],
-  [9.7749, 59.4224],
-  [9.7749, 59.4226],
-  [9.7748, 59.4229],
-  [9.7747, 59.4231],
-  [9.7746, 59.4234],
-  [9.7746, 59.4236],
-  [9.7745, 59.4238],
-  [9.7744, 59.4241],
-  [9.7743, 59.4243],
-  [9.7743, 59.4246],
-  [9.7742, 59.4248],
-  [9.7741, 59.4251],
-  [9.774, 59.4253],
-  [9.774, 59.4255],
-  [9.7739, 59.4258],
-  [9.7738, 59.426],
-  [9.7737, 59.4263],
-  [9.7737, 59.4266],
-  [9.7736, 59.4268],
-  [9.7736, 59.427],
-  [9.7735, 59.4273],
-  [9.7735, 59.4275],
-  [9.7734, 59.4278],
-  [9.7734, 59.428],
-  [9.7733, 59.4283],
-  [9.7733, 59.4285],
-  [9.7732, 59.4287],
-  [9.7732, 59.4288],
-  [9.7732, 59.4289],
-  [9.7732, 59.429],
-  [9.7731, 59.4293],
-  [9.7731, 59.4295],
-  [9.773, 59.4297],
-  [9.773, 59.43],
-  [9.7729, 59.4303],
-  [9.7729, 59.4305],
-  [9.7728, 59.4308],
-  [9.7728, 59.431],
-  [9.7727, 59.4312],
-  [9.7727, 59.4315],
-  [9.7726, 59.4317],
-  [9.7726, 59.432],
-  [9.7725, 59.4322],
-  [9.7724, 59.4325],
-  [9.7724, 59.4327],
-  [9.7723, 59.433],
-  [9.7723, 59.4332],
-  [9.7722, 59.4335],
-  [9.7722, 59.4337],
-  [9.7721, 59.434],
-  [9.7721, 59.4343],
-  [9.772, 59.4345],
-  [9.7719, 59.4348],
-  [9.7719, 59.4351],
-  [9.7718, 59.4353],
-  [9.7718, 59.4356],
-  [9.7717, 59.4359],
-  [9.7717, 59.4361],
-  [9.7716, 59.4364],
-  [9.7715, 59.4367],
-  [9.7715, 59.4369],
-  [9.7714, 59.4374],
-  [9.7712, 59.4376],
-  [9.7711, 59.4379],
-  [9.7709, 59.4381],
-  [9.7708, 59.4384],
-  [9.7706, 59.4386],
-  [9.7705, 59.4388],
-  [9.7703, 59.4391],
-  [9.7702, 59.4394],
-  [9.77, 59.4396],
-  [9.7699, 59.4399],
-  [9.7697, 59.4401],
-  [9.7695, 59.4404],
-  [9.7694, 59.4406],
-  [9.7692, 59.4409],
-  [9.7691, 59.4411],
-  [9.769, 59.4413],
-  [9.7688, 59.4416],
-  [9.7686, 59.4419],
-  [9.7684, 59.4422],
-  [9.7682, 59.4425],
-  [9.7681, 59.4428],
-  [9.7679, 59.4431],
-  [9.7677, 59.4433],
-  [9.7675, 59.4436],
-  [9.7673, 59.444],
-  [9.7671, 59.4443],
-  [9.767, 59.4446],
-  [9.7668, 59.4449],
-  [9.7666, 59.4452],
-  [9.7664, 59.4454],
-  [9.7662, 59.4457],
-  [9.7661, 59.446],
-  [9.7659, 59.4463],
-  [9.7657, 59.4466],
-  [9.7655, 59.4469],
-  [9.7653, 59.4472],
-  [9.7652, 59.4474],
-  [9.765, 59.4477],
-  [9.7648, 59.448],
-  [9.7647, 59.4483],
-  [9.7645, 59.4486],
-  [9.7643, 59.4489],
-  [9.7641, 59.4492],
-  [9.7639, 59.4495],
-  [9.7637, 59.4498],
-  [9.7636, 59.4501],
-  [9.7634, 59.4504],
-  [9.7632, 59.4506],
-  [9.763, 59.4509],
-  [9.7629, 59.4512],
-  [9.7627, 59.4515],
-  [9.7625, 59.4518],
-  [9.7624, 59.452],
-  [9.7622, 59.4523],
-  [9.762, 59.4526],
-  [9.7618, 59.4529],
-  [9.7616, 59.4532],
-  [9.7615, 59.4535],
-  [9.7613, 59.4538],
-  [9.7611, 59.4541],
-  [9.7609, 59.4543],
-  [9.7608, 59.4546],
-  [9.7606, 59.4549],
-  [9.7604, 59.4552],
-  [9.7602, 59.4555],
-  [9.76, 59.4558],
-  [9.7598, 59.4561],
-  [9.7597, 59.4564],
-  [9.7595, 59.4567],
-  [9.7593, 59.457],
-  [9.7591, 59.4573],
-  [9.7589, 59.4576],
-  [9.7589, 59.4576],
-  [9.7589, 59.4577],
-  [9.7588, 59.4578],
-  [9.7586, 59.458],
-  [9.7585, 59.4583],
-  [9.7588, 59.4582],
-  [9.7589, 59.4582],
-  [9.7592, 59.4582],
-  [9.7597, 59.4581],
-  [9.7603, 59.458],
-  [9.7609, 59.4579],
-  [9.7614, 59.4579],
-  [9.762, 59.4578],
-  [9.7625, 59.4577],
-  [9.7631, 59.4576],
-  [9.7637, 59.4575],
-  [9.7643, 59.4575],
-  [9.7649, 59.4574],
-  [9.7654, 59.4573],
-  [9.7659, 59.4572],
-  [9.7665, 59.4571],
-  [9.767, 59.4571],
-  [9.7672, 59.4571],
-  [9.7677, 59.4571],
-  [9.7683, 59.4571],
-  [9.7689, 59.4571],
-  [9.7694, 59.4571],
-  [9.77, 59.4571],
-  [9.7706, 59.457],
-  [9.7712, 59.457],
-  [9.7717, 59.457],
-  [9.7723, 59.457],
-  [9.7729, 59.457],
-  [9.7734, 59.457],
-  [9.774, 59.457],
-  [9.7746, 59.457],
-  [9.775, 59.457],
-  [9.7751, 59.457],
-  [9.7756, 59.4572],
-  [9.7761, 59.4573],
-  [9.7766, 59.4574],
-  [9.7771, 59.4576],
-  [9.7776, 59.4577],
-  [9.778, 59.4578],
-  [9.7785, 59.458],
-  [9.779, 59.4581],
-  [9.7795, 59.4582],
-  [9.7799, 59.4584],
-  [9.7805, 59.4585],
-  [9.781, 59.4587],
-  [9.7815, 59.4588],
-  [9.782, 59.4589],
-  [9.7825, 59.4591],
-  [9.783, 59.4592],
-  [9.7835, 59.4593],
-  [9.7839, 59.4595],
-  [9.7844, 59.4596],
-  [9.7849, 59.4597],
-  [9.7854, 59.4599],
-  [9.7859, 59.46],
-  [9.7864, 59.4602],
-  [9.7869, 59.4603],
-  [9.7874, 59.4605],
-  [9.788, 59.4606],
-  [9.7885, 59.4607],
-  [9.789, 59.4609],
-  [9.7895, 59.461],
-  [9.79, 59.4612],
-  [9.7901, 59.4612],
-  [9.7905, 59.4613],
-  [9.791, 59.4614],
-  [9.7916, 59.4616],
-  [9.7921, 59.4617],
-  [9.7926, 59.4619],
-  [9.7931, 59.462],
-  [9.7936, 59.4622],
-  [9.7941, 59.4623],
-  [9.7946, 59.4624],
-  [9.7952, 59.4626],
-  [9.7956, 59.4627],
-  [9.796, 59.4628],
-  [9.7963, 59.4629],
-  [9.7967, 59.463],
-  [9.7971, 59.4631],
-  [9.7975, 59.4632],
-  [9.7979, 59.4633],
-  [9.7983, 59.4635],
-  [9.7987, 59.4636],
-  [9.7991, 59.4637],
-  [9.7995, 59.4638],
-  [9.7999, 59.4639],
-  [9.8003, 59.464],
-  [9.8007, 59.4641],
-  [9.8011, 59.4642],
-  [9.8016, 59.4644],
-  [9.802, 59.4645],
-  [9.8024, 59.4646],
-  [9.8028, 59.4647],
-  [9.803, 59.4647],
-  [9.8032, 59.4648],
-  [9.8036, 59.4649],
-  [9.804, 59.465],
-  [9.8044, 59.4651],
-  [9.8048, 59.4653],
-  [9.8052, 59.4654],
-  [9.8056, 59.4655],
-  [9.806, 59.4656],
-  [9.8064, 59.4657],
-  [9.8068, 59.4658],
-  [9.8072, 59.4659],
-  [9.8076, 59.466],
-  [9.808, 59.4661],
-  [9.8084, 59.4662],
-  [9.8089, 59.4664],
-  [9.8093, 59.4665],
-  [9.8097, 59.4666],
-  [9.8101, 59.4667],
-  [9.8113, 59.467],
-  [9.8113, 59.467],
-  [9.8118, 59.4672],
-  [9.8121, 59.4673],
-  [9.8125, 59.4674],
-  [9.8128, 59.4674],
-  [9.813, 59.4675],
-  [9.8134, 59.4676],
-  [9.8138, 59.4677],
-  [9.8142, 59.4678],
-  [9.8146, 59.4679],
-  [9.815, 59.4681],
-  [9.8154, 59.4682],
-  [9.8157, 59.4683],
-  [9.816, 59.4683],
-  [9.8162, 59.4684],
-  [9.8166, 59.4685],
-  [9.817, 59.4686],
-  [9.8174, 59.4687],
-  [9.8178, 59.4688],
-  [9.8182, 59.4689],
-  [9.8186, 59.4691],
-  [9.8191, 59.4692],
-  [9.8191, 59.4692],
-  [9.8191, 59.4692],
-  [9.8193, 59.4691],
-  [9.8195, 59.469],
-  [9.8196, 59.469],
-  [9.8197, 59.4689],
-  [9.8198, 59.4688],
-  [9.8202, 59.4686],
-  [9.8203, 59.4686],
-  [9.8205, 59.4686],
-  [9.8207, 59.4685],
-  [9.8209, 59.4685],
-  [9.8213, 59.4685],
-  [9.8215, 59.4685],
-  [9.8218, 59.4685],
-  [9.8221, 59.4684],
-  [9.8223, 59.4684],
-  [9.8227, 59.4684],
-  [9.823, 59.4684],
-  [9.8232, 59.4684],
-  [9.8237, 59.4683],
-  [9.8241, 59.4682],
-  [9.8242, 59.4681],
-  [9.8247, 59.468],
-  [9.8249, 59.468],
-  [9.825, 59.4679],
-  [9.8253, 59.4679],
-  [9.8255, 59.468],
-  [9.8256, 59.468],
-  [9.8257, 59.468],
-  [9.826, 59.468],
-  [9.8265, 59.4681],
-  [9.8268, 59.4681],
-  [9.8276, 59.4683],
-  [9.8278, 59.4683],
-  [9.828, 59.4683],
-  [9.8282, 59.4683],
-  [9.8284, 59.4682],
-  [9.8287, 59.4682],
-  [9.8289, 59.4682],
-  [9.8291, 59.4682],
-  [9.8293, 59.4682],
-  [9.8295, 59.4681],
-  [9.8295, 59.4681],
-  [9.8296, 59.4681],
-  [9.8296, 59.468],
-  [9.8297, 59.468],
-  [9.8297, 59.468],
-  [9.8297, 59.468],
-  [9.8297, 59.4679],
-  [9.8297, 59.4679],
-  [9.8297, 59.4678],
-  [9.8297, 59.4678],
-  [9.8298, 59.4677],
-  [9.8298, 59.4677],
-  [9.8299, 59.4676],
-  [9.8299, 59.4676],
-  [9.83, 59.4676],
-  [9.83, 59.4676],
-  [9.8302, 59.4676],
-  [9.8303, 59.4676],
-  [9.8305, 59.4676],
-  [9.8305, 59.4676],
-  [9.8307, 59.4676],
-  [9.8309, 59.4676],
-  [9.831, 59.4676],
-  [9.8311, 59.4676],
-  [9.8312, 59.4677],
-  [9.8313, 59.4677],
-  [9.8314, 59.4677],
-  [9.8315, 59.4677],
-  [9.8317, 59.4677],
-  [9.8318, 59.4677],
-  [9.8319, 59.4677],
-  [9.8321, 59.4677],
-  [9.8322, 59.4677],
-  [9.8323, 59.4676],
-  [9.8324, 59.4675],
-  [9.8324, 59.4675],
-  [9.8326, 59.4674],
-  [9.8327, 59.4674],
-  [9.8327, 59.4673],
-  [9.8329, 59.4673],
-  [9.8329, 59.4672],
-  [9.8332, 59.467],
-  [9.8336, 59.4668],
-  [9.8337, 59.4668],
-  [9.8338, 59.4667],
-  [9.8339, 59.4667],
-  [9.8342, 59.4666],
-  [9.8343, 59.4665],
-  [9.8344, 59.4665],
-  [9.8346, 59.4665],
-  [9.8347, 59.4665],
-  [9.8348, 59.4665],
-  [9.8349, 59.4665],
-  [9.8349, 59.4664],
-  [9.8349, 59.4664],
-  [9.835, 59.4664],
-  [9.8351, 59.4663],
-  [9.8352, 59.4663],
-  [9.8353, 59.4662],
-  [9.8355, 59.4661],
-  [9.8358, 59.466],
-  [9.8359, 59.4658],
-  [9.8361, 59.4657],
-  [9.8366, 59.4653],
-  [9.8368, 59.4652],
-  [9.8371, 59.4651],
-  [9.8374, 59.465],
-  [9.8375, 59.465],
-  [9.8379, 59.4646],
-  [9.8381, 59.4645],
-  [9.8384, 59.4644],
-  [9.8385, 59.4644],
-  [9.8386, 59.4644],
-  [9.8387, 59.4644],
-  [9.8388, 59.4644],
-  [9.839, 59.4644],
-  [9.8393, 59.4644],
-  [9.8395, 59.4643],
-  [9.8396, 59.4643],
-  [9.8397, 59.4642],
-  [9.8398, 59.4641],
-  [9.8399, 59.4641],
-  [9.84, 59.464],
-  [9.8402, 59.4639],
-  [9.8404, 59.4638],
-  [9.8407, 59.4636],
-  [9.8408, 59.4635],
-  [9.841, 59.4633],
-  [9.8412, 59.4631],
-  [9.8412, 59.4631],
-  [9.8413, 59.463],
-  [9.8414, 59.463],
-  [9.8416, 59.463],
-  [9.8419, 59.463],
-  [9.842, 59.463],
-  [9.8421, 59.4631],
-  [9.8421, 59.4631],
-  [9.8421, 59.4632],
-  [9.8421, 59.4632],
-  [9.8423, 59.4634],
-  [9.8425, 59.4635],
-  [9.8426, 59.4635],
-  [9.8428, 59.4635],
-  [9.8429, 59.4635],
-  [9.843, 59.4635],
-  [9.8431, 59.4635],
-  [9.8431, 59.4635],
-  [9.8437, 59.4634],
-  [9.8441, 59.4634],
-  [9.8447, 59.4635],
-  [9.8452, 59.4636],
-  [9.8454, 59.4635],
-  [9.8457, 59.4634],
-  [9.8459, 59.4632],
-  [9.8461, 59.4631],
-  [9.8465, 59.4629],
-  [9.8468, 59.4628],
-  [9.8471, 59.4628],
-  [9.8476, 59.4626],
-  [9.8479, 59.4624],
-  [9.8482, 59.4622],
-  [9.8489, 59.4619],
-  [9.8495, 59.4615],
-  [9.8496, 59.4615],
-  [9.8508, 59.4605],
-  [9.8509, 59.4603],
-  [9.851, 59.4603],
-  [9.851, 59.4602],
-  [9.8517, 59.4599],
-  [9.8518, 59.4595],
-  [9.8517, 59.4592],
-  [9.8518, 59.459],
-  [9.852, 59.4588],
-  [9.8522, 59.4587],
-  [9.8523, 59.4587],
-  [9.8525, 59.4586],
-  [9.8529, 59.4583],
-  [9.8531, 59.4583],
-  [9.8534, 59.4584],
-  [9.8538, 59.4584],
-  [9.8543, 59.4584],
-  [9.8548, 59.4585],
-  [9.8553, 59.4585],
-  [9.8558, 59.4585],
-  [9.8562, 59.4586],
-  [9.8567, 59.4586],
-  [9.8572, 59.4586],
-  [9.8576, 59.4587],
-  [9.8581, 59.4587],
-  [9.8585, 59.4587],
-  [9.859, 59.4588],
-  [9.8595, 59.4588],
-  [9.8599, 59.4589],
-  [9.8604, 59.4589],
-  [9.8609, 59.4589],
-  [9.8614, 59.459],
-  [9.8615, 59.459],
-  [9.8618, 59.459],
-  [9.8623, 59.459],
-  [9.8627, 59.4591],
-  [9.8632, 59.4591],
-  [9.8637, 59.4591],
-  [9.8641, 59.4592],
-  [9.8646, 59.4592],
-  [9.865, 59.4592],
-  [9.8655, 59.4593],
-  [9.866, 59.4593],
-  [9.8664, 59.4593],
-  [9.8669, 59.4594],
-  [9.8674, 59.4594],
-  [9.8678, 59.4594],
-  [9.8682, 59.4595],
-  [9.8687, 59.4595],
-  [9.8691, 59.4595],
-  [9.8696, 59.4596],
-  [9.87, 59.4596],
-  [9.8705, 59.4596],
-  [9.8709, 59.4597],
-  [9.8713, 59.4597],
-  [9.8717, 59.4597],
-  [9.872, 59.4598],
-  [9.8725, 59.4598],
-  [9.873, 59.4598],
-  [9.8734, 59.4599],
-  [9.8739, 59.4599],
-  [9.8744, 59.4599],
-  [9.8748, 59.46],
-  [9.875, 59.4599],
-  [9.8755, 59.4599],
-  [9.8759, 59.4598],
-  [9.8764, 59.4598],
-  [9.877, 59.4597],
-  [9.8774, 59.4597],
-  [9.8779, 59.4596],
-  [9.8784, 59.4595],
-  [9.8789, 59.4594],
-  [9.8794, 59.4593],
-  [9.8795, 59.4593],
-  [9.8799, 59.4592],
-  [9.8804, 59.4591],
-  [9.8808, 59.4591],
-  [9.8812, 59.459],
-  [9.8817, 59.4589],
-  [9.8821, 59.4588],
-  [9.8826, 59.4587],
-  [9.883, 59.4587],
-  [9.8835, 59.4586],
-  [9.884, 59.4585],
-  [9.8844, 59.4584],
-  [9.8848, 59.4583],
-  [9.885, 59.4583],
-  [9.8854, 59.4582],
-  [9.8858, 59.4581],
-  [9.8859, 59.4581],
-  [9.8863, 59.458],
-  [9.8868, 59.458],
-  [9.8872, 59.4579],
-  [9.8877, 59.4579],
-  [9.8881, 59.4578],
-  [9.8886, 59.4578],
-  [9.8891, 59.4577],
-  [9.8895, 59.4577],
-  [9.8899, 59.4577],
-  [9.89, 59.4576],
-  [9.8905, 59.4576],
-  [9.8911, 59.4575],
-  [9.8916, 59.4575],
-  [9.892, 59.4575],
-  [9.8925, 59.4574],
-  [9.8929, 59.4574],
-  [9.8934, 59.4573],
-  [9.8939, 59.4573],
-  [9.8943, 59.4572],
-  [9.8947, 59.4572],
-  [9.8952, 59.4572],
-  [9.8956, 59.4571],
-  [9.8958, 59.4571],
-  [9.8963, 59.4571],
-  [9.8968, 59.457],
-  [9.8973, 59.457],
-  [9.8977, 59.4569],
-  [9.8982, 59.4569],
-  [9.8987, 59.4568],
-  [9.8991, 59.4568],
-  [9.8996, 59.4567],
-  [9.9001, 59.4567],
-  [9.9006, 59.4566],
-  [9.9011, 59.4566],
-  [9.9015, 59.4565],
-  [9.9019, 59.4565],
-  [9.902, 59.4565],
-  [9.9025, 59.4564],
-  [9.903, 59.4564],
-  [9.9035, 59.4563],
-  [9.904, 59.4563],
-  [9.9044, 59.4562],
-  [9.9049, 59.4562],
-  [9.9054, 59.4562],
-  [9.9057, 59.4562],
-  [9.9057, 59.4562],
-  [9.906, 59.4561],
-  [9.9065, 59.4561],
-  [9.907, 59.456],
-  [9.9074, 59.4559],
-  [9.9079, 59.4559],
-  [9.9082, 59.4558],
-  [9.9086, 59.4558],
-  [9.9091, 59.4558],
-  [9.9096, 59.4557],
-  [9.9101, 59.4557],
-  [9.9107, 59.4557],
-  [9.9112, 59.4557],
-  [9.9117, 59.4556],
-  [9.9122, 59.4556],
-  [9.9133, 59.4556],
-  [9.9138, 59.4556],
-  [9.9143, 59.4556],
-  [9.9148, 59.4556],
-  [9.9152, 59.4555],
-  [9.9157, 59.4555],
-  [9.916, 59.4555],
-  [9.9163, 59.4555],
-  [9.9168, 59.4555],
-  [9.9172, 59.4555],
-  [9.9177, 59.4555],
-  [9.9182, 59.4555],
-  [9.9187, 59.4555],
-  [9.9192, 59.4555],
-  [9.9197, 59.4555],
-  [9.9201, 59.4555],
-  [9.9201, 59.4555],
-  [9.9202, 59.4555],
-  [9.9204, 59.4555],
-  [9.9206, 59.4555],
-  [9.9209, 59.4557],
-  [9.921, 59.4557],
-  [9.9212, 59.4558],
-  [9.9214, 59.4559],
-  [9.9218, 59.456],
-  [9.9223, 59.4561],
-  [9.9227, 59.4562],
-  [9.9231, 59.4563],
-  [9.9236, 59.4564],
-  [9.924, 59.4565],
-  [9.9245, 59.4565],
-  [9.9249, 59.4566],
-  [9.9254, 59.4567],
-  [9.9258, 59.4568],
-  [9.9259, 59.4568],
-  [9.9263, 59.4569],
-  [9.9267, 59.4569],
-  [9.9268, 59.4569],
-  [9.9273, 59.4571],
-  [9.9278, 59.4572],
-  [9.9282, 59.4572],
-  [9.9282, 59.4573],
-  [9.9285, 59.4573],
-  [9.9287, 59.4574],
-  [9.9292, 59.4575],
-  [9.9296, 59.4576],
-  [9.9301, 59.4577],
-  [9.9302, 59.4577],
-  [9.9306, 59.4578],
-  [9.931, 59.4579],
-  [9.931, 59.4579],
-  [9.9311, 59.458],
-  [9.9315, 59.4581],
-  [9.9318, 59.4581],
-  [9.932, 59.4582],
-  [9.9324, 59.4583],
-  [9.9328, 59.4584],
-  [9.9333, 59.4585],
-  [9.9336, 59.4586],
-  [9.936, 59.4591],
-  [9.9368, 59.4593],
-  [9.9372, 59.4595],
-  [9.9376, 59.4598],
-  [9.9373, 59.4601],
-  [9.937, 59.4604],
-  [9.937, 59.4606],
-  [9.9369, 59.4609],
-  [9.9368, 59.4612],
-  [9.9367, 59.4617],
-  [9.9367, 59.462],
-  [9.9366, 59.4624],
-  [9.9366, 59.4626],
-  [9.9366, 59.463],
-  [9.9367, 59.4637],
-  [9.937, 59.4646],
-  [9.9379, 59.465],
-  [9.9381, 59.4658],
-  [9.9417, 59.4661],
-  [9.942, 59.4661],
-  [9.9425, 59.4662],
-  [9.9427, 59.4662],
-  [9.9431, 59.4663],
-  [9.9431, 59.4663],
-  [9.9448, 59.4683],
-  [9.9475, 59.4681],
-  [9.9479, 59.468],
-  [9.9508, 59.4679],
-  [9.9534, 59.4679],
-  [9.9537, 59.4679],
-  [9.9543, 59.4679],
-  [9.9569, 59.4681],
-  [9.9598, 59.468],
-  [9.9608, 59.468],
-  [9.9628, 59.468],
-  [9.9637, 59.4679],
-  [9.9638, 59.4678],
-  [9.9645, 59.4678],
-  [9.9645, 59.4678],
-  [9.9653, 59.4677],
-  [9.9664, 59.4676],
-  [9.9706, 59.4669],
-  [9.9723, 59.4666],
-  [9.9728, 59.4666],
-  [9.9742, 59.4664],
-  [9.9753, 59.4664],
-  [9.9756, 59.4664],
-  [9.9787, 59.466],
-  [9.9821, 59.4668],
-  [9.9826, 59.4669],
-  [9.9844, 59.4673],
-  [9.9851, 59.4675],
-  [9.987, 59.4677],
-  [9.9872, 59.4677],
-  [9.9872, 59.4677],
-  [9.9877, 59.4676],
-  [9.9891, 59.4676],
-  [9.9891, 59.4676],
-  [9.9922, 59.4674],
-  [9.9925, 59.4674],
-  [9.9957, 59.4671],
-  [9.9957, 59.4671],
-  [9.996, 59.4671],
-  [9.9965, 59.4671],
-  [9.9966, 59.4671],
-  [9.9977, 59.467],
-  [9.998, 59.467],
-  [9.9983, 59.467],
-  [9.9983, 59.467],
-  [9.9984, 59.4668],
-  [9.9991, 59.4641],
-  [9.9991, 59.4641],
-  [9.9991, 59.4638],
-  [9.9992, 59.4633],
-  [9.9994, 59.4628],
-  [9.9998, 59.4624],
-  [10, 59.462],
-  [10.0003, 59.4615],
-  [10.0007, 59.4611],
-  [10.0013, 59.4605],
-  [10.0023, 59.46],
-  [10.0033, 59.4596],
-  [10.0041, 59.4594],
-  [10.0052, 59.4592],
-  [10.0064, 59.4591],
-  [10.0077, 59.4591],
-  [10.0088, 59.4592],
-  [10.0098, 59.4591],
-  [10.0107, 59.4589],
-  [10.011, 59.4587],
-  [10.0111, 59.4585],
-  [10.0118, 59.4585],
-  [10.0119, 59.4585],
-  [10.012, 59.4585],
-  [10.0128, 59.4586],
-  [10.0138, 59.4587],
-  [10.0141, 59.4587],
-  [10.0152, 59.4588],
-  [10.0164, 59.4591],
-  [10.0174, 59.4593],
-  [10.0184, 59.4595],
-  [10.0185, 59.4596],
-  [10.0187, 59.4596],
-  [10.0194, 59.4597],
-  [10.019, 59.4602],
-  [10.0191, 59.4603],
-  [10.0195, 59.4604],
-  [10.0199, 59.4605],
-  [10.0208, 59.4606],
-  [10.0213, 59.4607],
-  [10.0218, 59.4608],
-  [10.0224, 59.4609],
-  [10.0219, 59.4614],
-  [10.0217, 59.4619],
-  [10.0214, 59.4621],
-  [10.0213, 59.4624],
-  [10.0217, 59.4627],
-  [10.0218, 59.4628],
-  [10.0218, 59.463],
-  [10.0219, 59.4633],
-  [10.022, 59.4635],
-  [10.0223, 59.464],
-  [10.0225, 59.4644],
-  [10.0229, 59.4647],
-  [10.024, 59.4648],
-  [10.0245, 59.465],
-  [10.0253, 59.465],
-  [10.0259, 59.465],
-  [10.0259, 59.465],
-  [10.0273, 59.4653],
-  [10.0278, 59.4655],
-  [10.0284, 59.4656],
-  [10.0287, 59.4658],
-  [10.0296, 59.4658],
-  [10.0299, 59.4659],
-  [10.0302, 59.466],
-  [10.0308, 59.4662],
-  [10.0314, 59.4663],
-  [10.0319, 59.4663],
-  [10.032, 59.4663],
-  [10.032, 59.4663],
-  [10.0321, 59.4663],
-  [10.0324, 59.4663],
-  [10.0334, 59.4664],
-  [10.0336, 59.4666],
-  [10.0337, 59.4667],
-  [10.0345, 59.4669],
-  [10.0361, 59.4671],
-  [10.037, 59.4672],
-  [10.0371, 59.4672],
-  [10.0376, 59.4672],
-  [10.0386, 59.4673],
-  [10.0397, 59.4674],
-  [10.0405, 59.4676],
-  [10.0426, 59.4681],
-  [10.0426, 59.4681],
-  [10.0458, 59.4689],
-  [10.0476, 59.4694],
-  [10.0499, 59.4701],
-  [10.0506, 59.4703],
-  [10.0508, 59.4704],
-  [10.0519, 59.471],
-  [10.0525, 59.4713],
-  [10.054, 59.4717],
-  [10.0563, 59.4723],
-  [10.0569, 59.4724],
-  [10.06, 59.473],
-  [10.0597, 59.4734],
-  [10.0594, 59.4736],
-  [10.0591, 59.4741],
-  [10.0586, 59.4745],
-  [10.0585, 59.4747],
-  [10.0582, 59.4751],
-  [10.0576, 59.4759],
-  [10.0573, 59.4765],
-  [10.0571, 59.4769],
-  [10.057, 59.4771],
-  [10.0566, 59.4773],
-  [10.0565, 59.4777],
-  [10.0561, 59.4783],
-  [10.0557, 59.4787],
-  [10.0556, 59.4789],
-  [10.0552, 59.4797],
-  [10.0551, 59.4799],
-  [10.0548, 59.4804],
-  [10.0541, 59.4817],
-  [10.0541, 59.4817],
-  [10.054, 59.4818],
-  [10.0538, 59.4822],
-  [10.0534, 59.4828],
-  [10.0532, 59.4833],
-  [10.0531, 59.4835],
-  [10.0526, 59.484],
-  [10.0526, 59.484],
-  [10.0525, 59.4841],
-  [10.0524, 59.4842],
-  [10.0522, 59.4843],
-  [10.0518, 59.4847],
-  [10.0506, 59.4857],
-  [10.0506, 59.4857],
-  [10.0504, 59.4859],
-  [10.0501, 59.4861],
-  [10.049, 59.4867],
-  [10.0486, 59.4872],
-  [10.0485, 59.4873],
-  [10.0476, 59.488],
-  [10.0471, 59.4885],
-  [10.047, 59.4885],
-  [10.0464, 59.4892],
-  [10.0461, 59.4894],
-  [10.0457, 59.4898],
-  [10.0455, 59.4901],
-  [10.0453, 59.4905],
-  [10.0451, 59.4909],
-  [10.0448, 59.4914],
-  [10.0447, 59.4918],
-  [10.0447, 59.492],
-  [10.0446, 59.4923],
-  [10.0444, 59.4927],
-  [10.0443, 59.4931],
-  [10.0442, 59.4934],
-  [10.0442, 59.4936],
-  [10.044, 59.494],
-  [10.0438, 59.4948],
-  [10.0438, 59.4948],
-  [10.0438, 59.4949],
-  [10.0438, 59.4949],
-  [10.0437, 59.495],
-  [10.0437, 59.495],
-  [10.0434, 59.4958],
-  [10.0434, 59.4958],
-  [10.0431, 59.4962],
-  [10.0422, 59.4969],
-  [10.0417, 59.4973],
-  [10.0408, 59.4981],
-  [10.0402, 59.4986],
-  [10.0397, 59.4991],
-  [10.0391, 59.4994],
-  [10.0388, 59.4996],
-  [10.0382, 59.5],
-  [10.0378, 59.5004],
-  [10.0374, 59.5007],
-  [10.0373, 59.5011],
-  [10.0369, 59.5016],
-  [10.0366, 59.502],
-  [10.0363, 59.5023],
-  [10.0361, 59.5027],
-  [10.036, 59.5029],
-  [10.0352, 59.5036],
-  [10.0347, 59.5039],
-  [10.034, 59.5045],
-  [10.0338, 59.5046],
-  [10.0331, 59.505],
-  [10.0324, 59.5056],
-  [10.0323, 59.5057],
-  [10.0316, 59.5063],
-  [10.0312, 59.5067],
-  [10.0304, 59.5074],
-  [10.0293, 59.5084],
-  [10.0283, 59.5093],
-  [10.0269, 59.5106],
-  [10.025, 59.5123],
-  [10.0225, 59.5136],
-  [10.0218, 59.514],
-  [10.0201, 59.5146],
-  [10.0178, 59.516],
-  [10.0168, 59.5166],
-  [10.0112, 59.5199],
-  [10.0107, 59.5202],
-  [10.01, 59.5205],
-  [10.0096, 59.5208],
-  [10.0093, 59.521],
-  [10.009, 59.5211],
-  [10.0088, 59.5213],
-  [10.0083, 59.5215],
-  [10.0078, 59.5218],
-  [10.0073, 59.5221],
-  [10.007, 59.5223],
-  [10.0067, 59.5224],
-  [10.0064, 59.5226],
-  [10.0061, 59.5228],
-  [10.0058, 59.5229],
-  [10.0056, 59.5231],
-  [10.0053, 59.5232],
-  [10.0048, 59.5234],
-  [10.0044, 59.5236],
-  [10.0039, 59.5238],
-  [10.0034, 59.524],
-  [10.0034, 59.524],
-  [10.0026, 59.5245],
-  [10.0021, 59.5249],
-  [10.0017, 59.5251],
-  [9.9999, 59.5263],
-  [9.9987, 59.5268],
-  [9.997, 59.5274],
-  [9.9961, 59.5278],
-  [9.9956, 59.5279],
-  [9.9947, 59.5284],
-  [9.9942, 59.5287],
-  [9.994, 59.5288],
-  [9.9939, 59.5288],
-  [9.9929, 59.529],
-  [9.9922, 59.5291],
-  [9.9919, 59.5293],
-  [9.9912, 59.5296],
-  [9.9905, 59.53],
-  [9.99, 59.5302],
-  [9.9896, 59.5303],
-  [9.9894, 59.5303],
-  [9.9893, 59.5304],
-  [9.9889, 59.5306],
-  [9.9884, 59.531],
-  [9.9875, 59.5317],
-  [9.9873, 59.5319],
-  [9.9867, 59.5324],
-  [9.986, 59.5329],
-  [9.9852, 59.5335],
-  [9.9847, 59.5339],
-  [9.9839, 59.5346],
-  [9.9835, 59.5348],
-  [9.9832, 59.5351],
-  [9.9829, 59.5353],
-  [9.9825, 59.5356],
-  [9.9821, 59.5358],
-  [9.9818, 59.5361],
-  [9.9813, 59.5365],
-  [9.9807, 59.5369],
-  [9.9802, 59.5374],
-  [9.9798, 59.5376],
-  [9.9793, 59.5379],
-  [9.9787, 59.5382],
-  [9.9781, 59.5386],
-  [9.9775, 59.539],
-  [9.9766, 59.5396],
-  [9.9757, 59.5401],
-  [9.9753, 59.5405],
-  [9.9748, 59.5411],
-  [9.9742, 59.5413],
-  [9.9739, 59.5416],
-  [9.9732, 59.5423],
-  [9.9727, 59.5428],
-  [9.9721, 59.5432],
-  [9.9716, 59.5435],
-  [9.9714, 59.5438],
-  [9.9708, 59.5443],
-  [9.9704, 59.5447],
-  [9.97, 59.545],
-  [9.9695, 59.5454],
-  [9.9691, 59.5458],
-  [9.9686, 59.5461],
-  [9.9681, 59.5463],
-  [9.9678, 59.5468],
-  [9.9674, 59.5473],
-  [9.9673, 59.5474],
-  [9.9669, 59.5477],
-  [9.9667, 59.5478],
-  [9.9664, 59.548],
-  [9.9659, 59.5484],
-  [9.9653, 59.5489],
-  [9.9648, 59.5493],
-  [9.9645, 59.5495],
-  [9.9643, 59.5498],
-  [9.964, 59.55],
-  [9.9639, 59.5501],
-  [9.9638, 59.5502],
-  [9.9633, 59.5506],
-  [9.963, 59.5508],
-  [9.9627, 59.5511],
-  [9.9626, 59.5511],
-  [9.9616, 59.552],
-  [9.9611, 59.5524],
-  [9.9607, 59.5528],
-  [9.9604, 59.553],
-  [9.9603, 59.5532],
-  [9.9586, 59.5546],
-  [9.9582, 59.5549],
-  [9.9579, 59.5551],
-  [9.9577, 59.5553],
-  [9.9575, 59.5555],
-  [9.9559, 59.557],
-  [9.9558, 59.5572],
-  [9.9555, 59.5574],
-  [9.9538, 59.5587],
-  [9.9534, 59.559],
-  [9.9532, 59.5591],
-  [9.9531, 59.5592],
-  [9.9525, 59.5598],
-  [9.9522, 59.5601],
-  [9.952, 59.5603],
-  [9.9518, 59.5605],
-  [9.9516, 59.5605],
-  [9.9516, 59.5606],
-  [9.9513, 59.5608],
-  [9.9512, 59.5609],
-  [9.951, 59.5611],
-  [9.9506, 59.5613],
-  [9.9504, 59.5615],
-  [9.9501, 59.5617],
-  [9.9499, 59.5619],
-  [9.9496, 59.5621],
-  [9.9493, 59.5623],
-  [9.949, 59.5625],
-  [9.9487, 59.5628],
-  [9.9485, 59.563],
-  [9.9481, 59.5634],
-  [9.9478, 59.5637],
-  [9.9477, 59.5638],
-  [9.9474, 59.564],
-  [9.9458, 59.5654],
-  [9.944, 59.5669],
-  [9.9431, 59.5676],
-  [9.9428, 59.5678],
-  [9.9425, 59.568],
-  [9.9422, 59.5682],
-  [9.9417, 59.5684],
-  [9.9414, 59.5686],
-  [9.9412, 59.5687],
-  [9.941, 59.5688],
-  [9.9406, 59.5691],
-  [9.9403, 59.5693],
-  [9.94, 59.5696],
-  [9.9398, 59.5698],
-  [9.9395, 59.57],
-  [9.9391, 59.5703],
-  [9.9388, 59.5705],
-  [9.9384, 59.5707],
-  [9.9381, 59.5709],
-  [9.9378, 59.571],
-  [9.9376, 59.5712],
-  [9.9375, 59.5712],
-  [9.9371, 59.5715],
-  [9.9366, 59.5718],
-  [9.936, 59.5722],
-  [9.9353, 59.5726],
-  [9.935, 59.5728],
-  [9.9342, 59.5732],
-  [9.9338, 59.5734],
-  [9.9335, 59.5736],
-  [9.9329, 59.574],
-  [9.9326, 59.5741],
-  [9.9322, 59.5743],
-  [9.9316, 59.5747],
-  [9.9313, 59.5749],
-  [9.9309, 59.5751],
-  [9.9306, 59.5753],
-  [9.9303, 59.5755],
-  [9.93, 59.5756],
-  [9.9298, 59.5757],
-  [9.9295, 59.5759],
-  [9.9292, 59.5761],
-  [9.9287, 59.5764],
-  [9.9284, 59.5766],
-  [9.9278, 59.5769],
-  [9.927, 59.5773],
-  [9.9263, 59.5776],
-  [9.9257, 59.578],
-  [9.9256, 59.578],
-  [9.9253, 59.5782],
-  [9.925, 59.5785],
-  [9.9246, 59.5788],
-  [9.9243, 59.579],
-  [9.9241, 59.5791],
-  [9.9236, 59.5793],
-  [9.9231, 59.5795],
-  [9.923, 59.5795],
-  [9.9222, 59.5798],
-  [9.9214, 59.5802],
-  [9.9208, 59.5806],
-  [9.9205, 59.5807],
-  [9.9199, 59.5809],
-  [9.9194, 59.5811],
-  [9.9189, 59.5813],
-  [9.9183, 59.5815],
-  [9.9178, 59.5816],
-  [9.9173, 59.5818],
-  [9.9166, 59.5821],
-  [9.9158, 59.5823],
-  [9.9153, 59.5825],
-  [9.9154, 59.5825],
-  [9.9157, 59.5825],
-  [9.9166, 59.5825],
-  [9.9176, 59.5825],
-  [9.9182, 59.5825],
-  [9.9186, 59.5825],
-  [9.9196, 59.5825],
-  [9.9204, 59.5825],
-  [9.9205, 59.5825],
-  [9.921, 59.5825],
-  [9.9211, 59.5825],
-  [9.9212, 59.5825],
-  [9.9217, 59.5828],
-  [9.9219, 59.5828],
-  [9.9224, 59.5831],
-  [9.923, 59.5834],
-  [9.923, 59.5834],
-  [9.9235, 59.5836],
-  [9.9239, 59.5839],
-  [9.9242, 59.5841],
-  [9.9246, 59.5843],
-  [9.925, 59.5845],
-  [9.9254, 59.5848],
-  [9.9256, 59.5849],
-  [9.9258, 59.585],
-  [9.9288, 59.5859],
-  [9.9292, 59.5859],
-  [9.9294, 59.586],
-  [9.9295, 59.586],
-  [9.9297, 59.5861],
-  [9.9297, 59.5861],
-  [9.9297, 59.5862],
-  [9.9297, 59.5863],
-  [9.9297, 59.5863],
-  [9.9298, 59.5863],
-  [9.9298, 59.5863],
-  [9.9299, 59.5863],
-  [9.93, 59.5863],
-  [9.9301, 59.5863],
-  [9.9302, 59.5863],
-  [9.9303, 59.5863],
-  [9.9304, 59.5863],
-  [9.9305, 59.5863],
-  [9.9306, 59.5864],
-  [9.9308, 59.5864],
-  [9.9309, 59.5864],
-  [9.9311, 59.5865],
-  [9.9311, 59.5865],
-  [9.9313, 59.5865],
-  [9.9313, 59.5865],
-  [9.9316, 59.5865],
-  [9.9316, 59.5865],
-  [9.9319, 59.5865],
-  [9.9319, 59.5865],
-  [9.9324, 59.5865],
-  [9.9351, 59.5865],
-  [9.9355, 59.5865],
-  [9.9418, 59.5869],
-  [9.9385, 59.5934],
-  [9.9383, 59.5939],
-  [9.9368, 59.5971],
-  [9.9367, 59.5974],
-  [9.9364, 59.5979],
-  [9.9349, 59.6],
-  [9.9347, 59.6003],
-  [9.9344, 59.6007],
-  [9.9332, 59.6029],
-  [9.933, 59.6033],
-  [9.9329, 59.6037],
-  [9.9329, 59.6039],
-  [9.9336, 59.6039],
-  [9.934, 59.6039],
-  [9.9341, 59.6039],
-  [9.9345, 59.6039],
-  [9.9352, 59.6042],
-  [9.9356, 59.6045],
-  [9.9358, 59.6046],
-  [9.9364, 59.6049],
-  [9.937, 59.6053],
-  [9.9375, 59.6057],
-  [9.9382, 59.6062],
-  [9.9388, 59.6066],
-  [9.9394, 59.6069],
-  [9.9395, 59.607],
-  [9.94, 59.6073],
-  [9.9403, 59.6075],
-  [9.9408, 59.6078],
-  [9.9413, 59.6082],
-  [9.9421, 59.6086],
-  [9.9429, 59.6089],
-  [9.9438, 59.6092],
-  [9.944, 59.6092],
-  [9.9446, 59.6094],
-  [9.9456, 59.6096],
-  [9.9464, 59.6098],
-  [9.9473, 59.61],
-  [9.9482, 59.6102],
-  [9.9491, 59.6103],
-  [9.95, 59.6104],
-  [9.9509, 59.6105],
-  [9.9519, 59.6107],
-  [9.9526, 59.6108],
-  [9.9533, 59.6111],
-  [9.9541, 59.6113],
-  [9.9548, 59.6115],
-  [9.955, 59.6115],
-  [9.9558, 59.6117],
-  [9.956, 59.6118],
-  [9.9567, 59.6119],
-  [9.9578, 59.6122],
-  [9.9587, 59.6124],
-  [9.9595, 59.6125],
-  [9.9602, 59.6128],
-  [9.961, 59.6131],
-  [9.9612, 59.6131],
-  [9.9619, 59.6134],
-  [9.9626, 59.6136],
-  [9.9634, 59.6139],
-  [9.9642, 59.6142],
-  [9.965, 59.6145],
-  [9.9659, 59.6148],
-  [9.9666, 59.615],
-  [9.9674, 59.6152],
-  [9.9683, 59.6154],
-  [9.9691, 59.6156],
-  [9.9697, 59.6157],
-  [9.9699, 59.6158],
-  [9.9708, 59.6159],
-  [9.9717, 59.6161],
-  [9.9726, 59.6163],
-  [9.9803, 59.6178],
-  [9.9811, 59.618],
-  [9.981, 59.618],
-  [9.9807, 59.6185],
-  [9.9771, 59.6231],
-  [9.9768, 59.6236],
-  [9.9765, 59.624],
-  [9.9762, 59.6243],
-  [9.9749, 59.6276],
-  [9.9748, 59.628],
-  [9.9747, 59.6285],
-  [9.9746, 59.6323],
-  [9.9746, 59.6325],
-  [9.9746, 59.6326],
-  [9.9747, 59.6327],
-  [9.9748, 59.6328],
-  [9.975, 59.6329],
-  [9.9751, 59.633],
-  [9.9751, 59.6331],
-  [9.9753, 59.6333],
-  [9.9756, 59.6336],
-  [9.9793, 59.6358],
-  [9.98, 59.6361],
-  [9.9809, 59.6365],
-  [9.9834, 59.6375],
-  [9.9861, 59.6387],
-  [9.9866, 59.639],
-  [9.9868, 59.6393],
-  [9.9868, 59.6397],
-  [9.9865, 59.6406],
-  [9.9858, 59.6414],
-  [9.9842, 59.6424],
-  [9.9842, 59.6424],
-  [9.9829, 59.6436],
-  [9.9837, 59.6441],
-  [9.9838, 59.6442],
-  [9.984, 59.6443],
-  [9.9841, 59.6445],
-  [9.9841, 59.6445],
-  [9.9842, 59.6447],
-  [9.9844, 59.6449],
-  [9.9846, 59.645],
-  [9.9846, 59.6451],
-  [9.9847, 59.6452],
-  [9.9847, 59.6452],
-  [9.9847, 59.6452],
-  [9.9847, 59.6452],
-  [9.9848, 59.6453],
-  [9.9848, 59.6454],
-  [9.9848, 59.6455],
-  [9.9849, 59.6455],
-  [9.9849, 59.6456],
-  [9.9849, 59.6457],
-  [9.985, 59.6457],
-  [9.985, 59.6458],
-  [9.9851, 59.6458],
-  [9.9853, 59.6459],
-  [9.9854, 59.6459],
-  [9.9856, 59.646],
-  [9.9857, 59.646],
-  [9.9859, 59.646],
-  [9.986, 59.6461],
-  [9.9862, 59.6461],
-  [9.9863, 59.6462],
-  [9.9864, 59.6462],
-  [9.9865, 59.6463],
-  [9.9866, 59.6463],
-  [9.9867, 59.6464],
-  [9.9869, 59.6465],
-  [9.9871, 59.6466],
-  [9.9873, 59.6467],
-  [9.9875, 59.6468],
-  [9.9876, 59.6469],
-  [9.9878, 59.6469],
-  [9.9879, 59.647],
-  [9.988, 59.6471],
-  [9.9881, 59.6472],
-  [9.9882, 59.6472],
-  [9.9883, 59.6473],
-  [9.9883, 59.6474],
-  [9.9883, 59.6475],
-  [9.9883, 59.6475],
-  [9.9883, 59.6476],
-  [9.9883, 59.6476],
-  [9.9884, 59.6477],
-  [9.9884, 59.6477],
-  [9.9885, 59.6478],
-  [9.9886, 59.6479],
-  [9.9887, 59.6479],
-  [9.9887, 59.648],
-  [9.9888, 59.6481],
-  [9.9889, 59.6481],
-  [9.989, 59.6482],
-  [9.9891, 59.6482],
-  [9.9892, 59.6483],
-  [9.9894, 59.6483],
-  [9.9895, 59.6483],
-  [9.9896, 59.6483],
-  [9.9897, 59.6484],
-  [9.9898, 59.6484],
-  [9.9898, 59.6485],
-  [9.9899, 59.6485],
-  [9.9899, 59.6486],
-  [9.9898, 59.6487],
-  [9.9898, 59.6487],
-  [9.9899, 59.6488],
-  [9.9899, 59.6488],
-  [9.99, 59.6489],
-  [9.9901, 59.6489],
-  [9.9902, 59.6489],
-  [9.9903, 59.6489],
-  [9.9905, 59.6489],
-  [9.9906, 59.6489],
-  [9.9907, 59.649],
-  [9.9909, 59.649],
-  [9.991, 59.649],
-  [9.9912, 59.6491],
-  [9.9914, 59.6491],
-  [9.9915, 59.6491],
-  [9.9916, 59.6492],
-  [9.9918, 59.6492],
-  [9.992, 59.6494],
-  [9.9921, 59.6495],
-  [9.9923, 59.6495],
-  [9.9924, 59.6496],
-  [9.9925, 59.6496],
-  [9.9926, 59.6497],
-  [9.9927, 59.6497],
-  [9.9928, 59.6497],
-  [9.9929, 59.6497],
-  [9.993, 59.6498],
-  [9.9931, 59.6498],
-  [9.9931, 59.6499],
-  [9.993, 59.65],
-  [9.9929, 59.65],
-  [9.9928, 59.6501],
-  [9.9928, 59.6502],
-  [9.9928, 59.6502],
-  [9.9929, 59.6503],
-  [9.9931, 59.6503],
-  [9.9932, 59.6504],
-  [9.9933, 59.6504],
-  [9.9934, 59.6504],
-  [9.9935, 59.6505],
-  [9.9935, 59.6505],
-  [9.9936, 59.6506],
-  [9.9936, 59.6507],
-  [9.9935, 59.6508],
-  [9.9935, 59.6508],
-  [9.9935, 59.6509],
-  [9.9935, 59.651],
-  [9.9935, 59.651],
-  [9.9937, 59.6511],
-  [9.9938, 59.6511],
-  [9.9939, 59.6512],
-  [9.994, 59.6512],
-  [9.994, 59.6513],
-  [9.9941, 59.6514],
-  [9.9941, 59.6514],
-  [9.9942, 59.6515],
-  [9.9942, 59.6516],
-  [9.9943, 59.6516],
-  [9.9944, 59.6517],
-  [9.9944, 59.6517],
-  [9.9945, 59.6517],
-  [9.9945, 59.6518],
-  [9.9946, 59.6519],
-  [9.9947, 59.652],
-  [9.9948, 59.6521],
-  [9.9949, 59.6522],
-  [9.995, 59.6524],
-  [9.9951, 59.6525],
-  [9.9951, 59.6526],
-  [9.9951, 59.6526],
-  [9.9951, 59.6527],
-  [9.9951, 59.6527],
-  [9.9951, 59.6528],
-  [9.9952, 59.6529],
-  [9.9952, 59.6529],
-  [9.9952, 59.653],
-  [9.9953, 59.6531],
-  [9.9954, 59.6532],
-  [9.9954, 59.6533],
-  [9.9954, 59.6533],
-  [9.9954, 59.6534],
-  [9.9953, 59.6535],
-  [9.9953, 59.6535],
-  [9.9954, 59.6536],
-  [9.9954, 59.6537],
-  [9.9954, 59.6537],
-  [9.9954, 59.6538],
-  [9.9953, 59.6538],
-  [9.9952, 59.6539],
-  [9.9951, 59.6539],
-  [9.9951, 59.654],
-  [9.995, 59.654],
-  [9.995, 59.6541],
-  [9.9951, 59.6541],
-  [9.9951, 59.6542],
-  [9.9952, 59.6542],
-  [9.9953, 59.6542],
-  [9.9954, 59.6542],
-  [9.9955, 59.6543],
-  [9.9956, 59.6543],
-  [9.9957, 59.6543],
-  [9.9958, 59.6544],
-  [9.9959, 59.6544],
-  [9.996, 59.6544],
-  [9.9961, 59.6545],
-  [9.9962, 59.6545],
-  [9.9963, 59.6545],
-  [9.9963, 59.6546],
-  [9.9964, 59.6546],
-  [9.9964, 59.6547],
-  [9.9964, 59.6547],
-  [9.9963, 59.6548],
-  [9.9963, 59.6549],
-  [9.9962, 59.655],
-  [9.9961, 59.655],
-  [9.996, 59.6551],
-  [9.9959, 59.6551],
-  [9.9958, 59.6552],
-  [9.9958, 59.6552],
-  [9.9958, 59.6553],
-  [9.9957, 59.6554],
-  [9.9957, 59.6555],
-  [9.9957, 59.6556],
-  [9.9957, 59.6556],
-  [9.9956, 59.6557],
-  [9.9955, 59.6557],
-  [9.9954, 59.6558],
-  [9.9954, 59.6559],
-  [9.9954, 59.656],
-  [9.9954, 59.656],
-  [9.9954, 59.6561],
-  [9.9954, 59.6561],
-  [9.9955, 59.6562],
-  [9.9956, 59.6562],
-  [9.9957, 59.6562],
-  [9.9959, 59.6562],
-  [9.9961, 59.6562],
-  [9.9962, 59.6562],
-  [9.9963, 59.6562],
-  [9.9964, 59.6561],
-  [9.9964, 59.6561],
-  [9.9965, 59.6561],
-  [9.9966, 59.6561],
-  [9.9966, 59.6561],
-  [9.9967, 59.6561],
-  [9.9968, 59.6561],
-  [9.997, 59.6562],
-  [9.9972, 59.6562],
-  [9.9974, 59.6562],
-  [9.9975, 59.6563],
-  [9.9977, 59.6563],
-  [9.9978, 59.6564],
-  [9.9979, 59.6564],
-  [9.998, 59.6564],
-  [9.9981, 59.6564],
-  [9.9982, 59.6564],
-  [9.9983, 59.6564],
-  [9.9984, 59.6564],
-  [9.9985, 59.6564],
-  [9.9987, 59.6564],
-  [9.999, 59.6564],
-  [9.9993, 59.6564],
-  [9.9995, 59.6564],
-  [9.9997, 59.6564],
-  [9.9998, 59.6565],
-  [9.9998, 59.6565],
-  [9.9998, 59.6565],
-  [9.9998, 59.6566],
-  [9.9998, 59.6567],
-  [9.9998, 59.6567],
-  [9.9997, 59.6568],
-  [9.9996, 59.6569],
-  [9.9995, 59.6569],
-  [9.9994, 59.657],
-  [9.9994, 59.657],
-  [9.9994, 59.657],
-  [9.9995, 59.6571],
-  [9.9995, 59.6572],
-  [9.9996, 59.6572],
-  [9.9997, 59.6573],
-  [9.9998, 59.6574],
-  [9.9999, 59.6575],
-  [10, 59.6575],
-  [10, 59.6576],
-  [10, 59.6576],
-  [9.9999, 59.6577],
-  [9.9998, 59.6577],
-  [9.9997, 59.6577],
-  [9.9996, 59.6577],
-  [9.9995, 59.6578],
-  [9.9995, 59.6578],
-  [9.9994, 59.6579],
-  [9.9994, 59.658],
-  [9.9994, 59.6581],
-  [9.9993, 59.6581],
-  [9.9993, 59.6582],
-  [9.9994, 59.6583],
-  [9.9994, 59.6584],
-  [9.9994, 59.6585],
-  [9.9995, 59.6586],
-  [9.9995, 59.6586],
-  [9.9995, 59.6587],
-  [9.9995, 59.6587],
-  [9.9994, 59.6588],
-  [9.9994, 59.6588],
-  [9.9993, 59.6589],
-  [9.9992, 59.6589],
-  [9.9991, 59.659],
-  [9.999, 59.6591],
-  [9.9989, 59.6592],
-  [9.9988, 59.6594],
-  [9.9987, 59.6596],
-  [9.9986, 59.6596],
-  [9.9985, 59.6597],
-  [9.9984, 59.6597],
-  [9.9984, 59.6598],
-  [9.9984, 59.6598],
-  [9.9984, 59.6599],
-  [9.9983, 59.66],
-  [9.9982, 59.6601],
-  [9.9982, 59.6601],
-  [9.9981, 59.6601],
-  [9.9981, 59.6602],
-  [9.9981, 59.6602],
-  [9.9982, 59.6602],
-  [9.9983, 59.6602],
-  [9.9984, 59.6603],
-  [9.9985, 59.6603],
-  [9.9986, 59.6604],
-  [9.9987, 59.6604],
-  [9.9988, 59.6605],
-  [9.9989, 59.6605],
-  [9.9991, 59.6605],
-  [9.9992, 59.6606],
-  [9.9993, 59.6606],
-  [9.9994, 59.6606],
-  [9.9995, 59.6606],
-  [9.9996, 59.6606],
-  [9.9997, 59.6607],
-  [9.9997, 59.6607],
-  [9.9997, 59.6607],
-  [9.9998, 59.6608],
-  [9.9998, 59.6608],
-  [9.9998, 59.6609],
-  [9.9998, 59.6609],
-  [9.9998, 59.661],
-  [9.9998, 59.6611],
-  [9.9998, 59.6613],
-  [9.9998, 59.6614],
-  [9.9998, 59.6615],
-  [9.9998, 59.6617],
-  [9.9998, 59.6618],
-  [9.9998, 59.6618],
-  [9.9997, 59.6619],
-  [9.9996, 59.662],
-  [9.9996, 59.6621],
-  [9.9996, 59.6621],
-  [9.9996, 59.6622],
-  [9.9996, 59.6623],
-  [9.9996, 59.6624],
-  [9.9996, 59.6624],
-  [9.9995, 59.6624],
-  [9.9994, 59.6624],
-  [9.9993, 59.6625],
-  [9.9993, 59.6625],
-  [9.9992, 59.6625],
-  [9.9992, 59.6626],
-  [9.9991, 59.6626],
-  [9.9991, 59.6627],
-  [9.999, 59.6628],
-  [9.999, 59.6628],
-  [9.9989, 59.6629],
-  [9.9989, 59.663],
-  [9.9988, 59.6631],
-  [9.9988, 59.6631],
-  [9.9988, 59.6632],
-  [9.9987, 59.6632],
-  [9.9987, 59.6632],
-  [9.9986, 59.6633],
-  [9.9986, 59.6633],
-  [9.9986, 59.6634],
-  [9.9985, 59.6634],
-  [9.9985, 59.6635],
-  [9.9985, 59.6635],
-  [9.9984, 59.6636],
-  [9.9983, 59.6636],
-  [9.9983, 59.6637],
-  [9.9982, 59.6637],
-  [9.9982, 59.6638],
-  [9.9982, 59.6638],
-  [9.9982, 59.6638],
-  [9.9982, 59.6638],
-  [9.9983, 59.6638],
-  [9.9983, 59.6639],
-  [9.9984, 59.6639],
-  [9.9985, 59.6639],
-  [9.9986, 59.6639],
-  [9.9986, 59.6639],
-  [9.9987, 59.6639],
-  [9.9988, 59.6639],
-  [9.9989, 59.6639],
-  [9.9991, 59.664],
-  [9.9992, 59.664],
-  [9.9994, 59.664],
-  [9.9996, 59.664],
-  [9.9997, 59.6639],
-  [9.9998, 59.6639],
-  [9.9999, 59.6639],
-  [10, 59.664],
-  [10.0002, 59.664],
-  [10.0003, 59.6641],
-  [10.0005, 59.6641],
-  [10.0006, 59.6641],
-  [10.0006, 59.6641],
-  [10.0007, 59.6641],
-  [10.0008, 59.6641],
-  [10.0009, 59.6641],
-  [10.001, 59.6641],
-  [10.0012, 59.6641],
-  [10.0014, 59.6641],
-  [10.0014, 59.6641],
-  [10.0015, 59.6641],
-  [10.0017, 59.6641],
-  [10.0018, 59.6642],
-  [10.002, 59.6642],
-  [10.0022, 59.6642],
-  [10.0022, 59.6642],
-  [10.0024, 59.6642],
-  [10.0026, 59.6643],
-  [10.0027, 59.6642],
-  [10.0029, 59.6642],
-  [10.003, 59.6642],
-  [10.0031, 59.6642],
-  [10.0032, 59.6642],
-  [10.0032, 59.6642],
-  [10.0034, 59.6643],
-  [10.0036, 59.6643],
-  [10.0037, 59.6644],
-  [10.0039, 59.6644],
-  [10.004, 59.6644],
-  [10.0042, 59.6645],
-  [10.0044, 59.6645],
-  [10.0045, 59.6645],
-  [10.0045, 59.6645],
-  [10.0046, 59.6645],
-  [10.0046, 59.6645],
-  [10.0047, 59.6644],
-  [10.0048, 59.6644],
-  [10.0049, 59.6644],
-  [10.005, 59.6643],
-  [10.0051, 59.6643],
-  [10.0051, 59.6642],
-  [10.0051, 59.6642],
-  [10.0052, 59.6642],
-  [10.0055, 59.6642],
-  [10.0053, 59.6642],
-  [10.0053, 59.6642],
-  [10.0054, 59.6641],
-  [10.0054, 59.6641],
-  [10.0054, 59.6641],
-  [10.0055, 59.664],
-  [10.0055, 59.664],
-  [10.0056, 59.664],
-  [10.0056, 59.6639],
-  [10.0057, 59.6639],
-  [10.0057, 59.6639],
-  [10.0057, 59.6639],
-  [10.0058, 59.6638],
-  [10.0058, 59.6638],
-  [10.0059, 59.6638],
-  [10.0059, 59.6638],
-  [10.0059, 59.6637],
-  [10.0059, 59.6637],
-  [10.006, 59.6637],
-  [10.006, 59.6637],
-  [10.006, 59.6636],
-  [10.0061, 59.6636],
-  [10.0061, 59.6636],
-  [10.0062, 59.6635],
-  [10.0062, 59.6635],
-  [10.0062, 59.6635],
-  [10.0063, 59.6635],
-  [10.0063, 59.6634],
-  [10.0063, 59.6634],
-  [10.0064, 59.6634],
-  [10.0064, 59.6634],
-  [10.0065, 59.6633],
-  [10.0065, 59.6633],
-  [10.0066, 59.6633],
-  [10.0066, 59.6632],
-  [10.0066, 59.6632],
-  [10.0066, 59.6632],
-  [10.0067, 59.6631],
-  [10.0067, 59.6631],
-  [10.0068, 59.6631],
-  [10.0068, 59.6631],
-  [10.0069, 59.663],
-  [10.0069, 59.663],
-  [10.0069, 59.663],
-  [10.007, 59.663],
-  [10.007, 59.6629],
-  [10.0071, 59.6629],
-  [10.0071, 59.6629],
-  [10.0072, 59.6628],
-  [10.0073, 59.6628],
-  [10.0073, 59.6628],
-  [10.0074, 59.6628],
-  [10.0074, 59.6627],
-  [10.0075, 59.6627],
-  [10.0076, 59.6626],
-  [10.0077, 59.6626],
-  [10.0077, 59.6626],
-  [10.0078, 59.6625],
-  [10.0078, 59.6625],
-  [10.0079, 59.6625],
-  [10.008, 59.6624],
-  [10.008, 59.6624],
-  [10.0081, 59.6623],
-  [10.0082, 59.6623],
-  [10.0083, 59.6623],
-  [10.0083, 59.6622],
-  [10.0084, 59.6622],
-  [10.0085, 59.6622],
-  [10.0085, 59.6622],
-  [10.0086, 59.6621],
-  [10.0087, 59.662],
-  [10.0088, 59.662],
-  [10.0089, 59.662],
-  [10.009, 59.6619],
-  [10.0091, 59.6619],
-  [10.0092, 59.6618],
-  [10.0094, 59.6618],
-  [10.0095, 59.6617],
-  [10.0095, 59.6617],
-  [10.0097, 59.6616],
-  [10.0098, 59.6616],
-  [10.0099, 59.6615],
-  [10.01, 59.6615],
-  [10.0102, 59.6614],
-  [10.0103, 59.6614],
-  [10.0104, 59.6613],
-  [10.0105, 59.6612],
-  [10.0107, 59.6612],
-  [10.0108, 59.6611],
-  [10.0108, 59.6611],
-  [10.0109, 59.6611],
-  [10.011, 59.6611],
-  [10.011, 59.6611],
-  [10.0111, 59.661],
-  [10.0112, 59.661],
-  [10.0112, 59.6609],
-  [10.0113, 59.6609],
-  [10.0113, 59.6609],
-  [10.0114, 59.6608],
-  [10.0114, 59.6608],
-  [10.0115, 59.6607],
-  [10.0116, 59.6607],
-  [10.0117, 59.6606],
-  [10.0117, 59.6606],
-  [10.0118, 59.6605],
-  [10.0118, 59.6605],
-  [10.0118, 59.6605],
-  [10.012, 59.6604],
-  [10.012, 59.6604],
-  [10.0121, 59.6603],
-  [10.0121, 59.6603],
-  [10.0122, 59.6603],
-  [10.0123, 59.6602],
-  [10.0123, 59.6602],
-  [10.0124, 59.6602],
-  [10.0125, 59.6601],
-  [10.0125, 59.6601],
-  [10.0126, 59.66],
-  [10.0127, 59.66],
-  [10.0128, 59.6599],
-  [10.0128, 59.6599],
-  [10.0129, 59.6599],
-  [10.0129, 59.6599],
-  [10.013, 59.6598],
-  [10.0131, 59.6598],
-  [10.0131, 59.6598],
-  [10.0132, 59.6598],
-  [10.0132, 59.6597],
-  [10.0133, 59.6597],
-  [10.0135, 59.6596],
-  [10.0135, 59.6596],
-  [10.0137, 59.6596],
-  [10.0138, 59.6595],
-  [10.0139, 59.6595],
-  [10.0139, 59.6595],
-  [10.0141, 59.6594],
-  [10.0142, 59.6593],
-  [10.0143, 59.6593],
-  [10.0144, 59.6593],
-  [10.0144, 59.6593],
-  [10.0145, 59.6593],
-  [10.0146, 59.6592],
-  [10.0147, 59.6592],
-  [10.0147, 59.6592],
-  [10.0148, 59.6592],
-  [10.0148, 59.6591],
-  [10.0149, 59.6591],
-  [10.015, 59.659],
-  [10.015, 59.659],
-  [10.0151, 59.659],
-  [10.0151, 59.6589],
-  [10.0152, 59.6589],
-  [10.0152, 59.6589],
-  [10.0153, 59.6588],
-  [10.0154, 59.6588],
-  [10.0155, 59.6588],
-  [10.0156, 59.6587],
-  [10.0157, 59.6587],
-  [10.0157, 59.6587],
-  [10.0158, 59.6587],
-  [10.0158, 59.6587],
-  [10.0159, 59.6587],
-  [10.016, 59.6586],
-  [10.016, 59.6586],
-  [10.0161, 59.6587],
-  [10.0162, 59.6587],
-  [10.0163, 59.6587],
-  [10.0163, 59.6587],
-  [10.0165, 59.6587],
-  [10.0165, 59.6587],
-  [10.0166, 59.6586],
-  [10.0166, 59.6586],
-  [10.0166, 59.6586],
-  [10.0166, 59.6585],
-  [10.0167, 59.6585],
-  [10.0167, 59.6585],
-  [10.0168, 59.6585],
-  [10.0168, 59.6585],
-  [10.0169, 59.6585],
-  [10.0169, 59.6584],
-  [10.017, 59.6584],
-  [10.017, 59.6584],
-  [10.0171, 59.6583],
-  [10.0171, 59.6583],
-  [10.0172, 59.6584],
-  [10.0172, 59.6584],
-  [10.0172, 59.6584],
-  [10.0173, 59.6585],
-  [10.0174, 59.6585],
-  [10.0174, 59.6585],
-  [10.0175, 59.6585],
-  [10.0176, 59.6585],
-  [10.0177, 59.6584],
-  [10.0178, 59.6584],
-  [10.0179, 59.6584],
-  [10.018, 59.6584],
-  [10.0181, 59.6584],
-  [10.0182, 59.6584],
-  [10.0183, 59.6583],
-  [10.0184, 59.6583],
-  [10.0185, 59.6583],
-  [10.0186, 59.6582],
-  [10.0187, 59.6582],
-  [10.0187, 59.6582],
-  [10.0188, 59.6582],
-  [10.0188, 59.6581],
-  [10.0188, 59.6581],
-  [10.0188, 59.6581],
-  [10.0189, 59.6581],
-  [10.019, 59.6581],
-  [10.019, 59.658],
-  [10.0191, 59.658],
-  [10.0191, 59.658],
-  [10.0192, 59.658],
-  [10.0192, 59.658],
-  [10.0193, 59.6579],
-  [10.0193, 59.6579],
-  [10.0194, 59.6579],
-  [10.0194, 59.6579],
-  [10.0194, 59.6578],
-  [10.0194, 59.6578],
-  [10.0194, 59.6578],
-  [10.0194, 59.6578],
-  [10.0198, 59.6575],
-  [10.0218, 59.6553],
-  [10.0217, 59.6552],
-  [10.0223, 59.655],
-  [10.0228, 59.6548],
-  [10.0234, 59.6545],
-  [10.0237, 59.6545],
-  [10.0243, 59.6541],
-  [10.0248, 59.6539],
-  [10.0255, 59.6534],
-  [10.026, 59.653],
-  [10.0265, 59.6525],
-  [10.0268, 59.6522],
-  [10.0272, 59.6519],
-  [10.0277, 59.6514],
-  [10.0281, 59.6512],
-  [10.0289, 59.6502],
-  [10.0292, 59.65],
-  [10.0297, 59.6494],
-  [10.0304, 59.6488],
-  [10.0323, 59.647],
-  [10.0328, 59.6467],
-  [10.0333, 59.6458],
-  [10.0333, 59.6458],
-  [10.0333, 59.6457],
-  [10.0333, 59.6456],
-  [10.0333, 59.6455],
-  [10.0334, 59.6454],
-  [10.0334, 59.6453],
-  [10.0335, 59.6452],
-  [10.0335, 59.6452],
-  [10.0335, 59.6451],
-  [10.0336, 59.645],
-  [10.0337, 59.645],
-  [10.0337, 59.6449],
-  [10.0338, 59.6448],
-  [10.0339, 59.6447],
-  [10.034, 59.6446],
-  [10.0341, 59.6445],
-  [10.0341, 59.6444],
-  [10.0342, 59.6444],
-  [10.0342, 59.6444],
-  [10.0343, 59.6442],
-  [10.0344, 59.6441],
-  [10.0345, 59.644],
-  [10.0346, 59.6439],
-  [10.0347, 59.6438],
-  [10.0348, 59.6438],
-  [10.0349, 59.6438],
-  [10.035, 59.6437],
-  [10.0351, 59.6436],
-  [10.0352, 59.6436],
-  [10.0353, 59.6435],
-  [10.0354, 59.6434],
-  [10.0355, 59.6434],
-  [10.0356, 59.6433],
-  [10.0356, 59.6432],
-  [10.0357, 59.6432],
-  [10.0357, 59.6432],
-  [10.0357, 59.6432],
-  [10.0358, 59.6431],
-  [10.0359, 59.643],
-  [10.036, 59.6429],
-  [10.036, 59.6428],
-  [10.0361, 59.6427],
-  [10.0362, 59.6427],
-  [10.0362, 59.6426],
-  [10.0363, 59.6426],
-  [10.0363, 59.6426],
-  [10.0363, 59.6426],
-  [10.0363, 59.6426],
-  [10.0373, 59.6408],
-  [10.0375, 59.6405],
-  [10.0382, 59.6406],
-  [10.0383, 59.6406],
-  [10.0445, 59.6415],
-  [10.0474, 59.6433],
-  [10.0522, 59.6439],
-  [10.0608, 59.6437],
-  [10.0699, 59.6435],
-  [10.0778, 59.6432],
-  [10.0889, 59.6426],
-  [10.0919, 59.6472],
-  [10.0949, 59.6512],
-  [10.0986, 59.6561],
-  [10.1013, 59.6581],
-  [10.1061, 59.6591],
-  [10.1086, 59.6596],
-  [10.1106, 59.6602],
-  [10.1115, 59.6607],
-  [10.1119, 59.6614],
-  [10.1122, 59.662],
-  [10.1123, 59.6625],
-  [10.1123, 59.6626],
-  [10.1124, 59.6629],
-  [10.1124, 59.663],
-  [10.1126, 59.6635],
-  [10.1126, 59.6636],
-  [10.1124, 59.6638],
-  [10.1124, 59.6638],
-  [10.1124, 59.6639],
-  [10.1124, 59.664],
-  [10.1125, 59.6643],
-  [10.1128, 59.6648],
-  [10.1128, 59.6651],
-  [10.113, 59.6659],
-  [10.1128, 59.666],
-  [10.1133, 59.6668],
-  [10.1132, 59.6669],
-  [10.1134, 59.6674],
-  [10.1127, 59.6678],
-  [10.1145, 59.6683],
-  [10.115, 59.6691],
-  [10.1155, 59.6707],
-  [10.1239, 59.6712],
-  [10.1288, 59.6711],
-  [10.1362, 59.6717],
-  [10.1432, 59.6691],
-  [10.1438, 59.6691],
-  [10.1455, 59.6689],
-  [10.1467, 59.6687],
-  [10.147, 59.6685],
-  [10.1503, 59.6678],
-  [10.1525, 59.6675],
-  [10.1533, 59.6673],
-  [10.1539, 59.6672],
-  [10.1545, 59.6671],
-  [10.1559, 59.6666],
-  [10.1568, 59.666],
-  [10.1572, 59.6658],
-  [10.1578, 59.6663],
-  [10.1581, 59.6667],
-  [10.1574, 59.6677],
-  [10.1574, 59.6677],
-  [10.1572, 59.668],
-  [10.1582, 59.6683],
-  [10.1585, 59.6684],
-  [10.1587, 59.6686],
-  [10.159, 59.6688],
-  [10.1594, 59.6689],
-  [10.1596, 59.669],
-  [10.1602, 59.669],
-  [10.1602, 59.6691],
-  [10.1611, 59.669],
-  [10.1619, 59.6687],
-  [10.1626, 59.6684],
-  [10.1637, 59.6682],
-  [10.1647, 59.6681],
-  [10.1653, 59.668],
-  [10.1663, 59.6678],
-  [10.1706, 59.6683],
-  [10.171, 59.6684],
-  [10.1711, 59.6684],
-  [10.1712, 59.6684],
-  [10.1713, 59.6684],
-  [10.1716, 59.6685],
-  [10.1724, 59.6687],
-  [10.1739, 59.6692],
-  [10.1742, 59.6694],
-  [10.1747, 59.6692],
-  [10.1753, 59.669],
-  [10.1758, 59.6688],
-  [10.1765, 59.6687],
-  [10.1774, 59.6685],
-  [10.1779, 59.6683],
-  [10.1784, 59.6681],
-  [10.1797, 59.6681],
-  [10.1798, 59.668],
-  [10.1802, 59.6681],
-  [10.1804, 59.6681],
-  [10.1816, 59.6683],
-  [10.1828, 59.6684],
-  [10.1836, 59.6685],
-  [10.1845, 59.6686],
-  [10.1847, 59.6686],
-  [10.1855, 59.6688],
-  [10.1859, 59.6686],
-  [10.1862, 59.6686],
-  [10.1866, 59.6689],
-  [10.1864, 59.669],
-  [10.1866, 59.6693],
-  [10.1869, 59.6696],
-  [10.1868, 59.6697],
-  [10.187, 59.6698],
-  [10.1875, 59.6702],
-  [10.1877, 59.6703],
-  [10.1878, 59.6705],
-  [10.1882, 59.6706],
-  [10.1886, 59.6706],
-  [10.1894, 59.6709],
-  [10.1899, 59.6709],
-  [10.1904, 59.6709],
-  [10.1908, 59.6709],
-  [10.1913, 59.6709],
-  [10.1915, 59.6711],
-  [10.1916, 59.6713],
-  [10.1917, 59.6715],
-  [10.1917, 59.6716],
-  [10.1922, 59.6718],
-  [10.1924, 59.6719],
-  [10.1927, 59.6719],
-  [10.1928, 59.6719],
-  [10.1929, 59.6718],
-  [10.193, 59.6717],
-  [10.193, 59.6717],
-  [10.1931, 59.6716],
-  [10.1932, 59.6716],
-  [10.1938, 59.6715],
-  [10.1953, 59.6714],
-  [10.1961, 59.6713],
-  [10.1965, 59.6714],
-  [10.1968, 59.6713],
-  [10.1976, 59.671],
-  [10.1981, 59.6708],
-  [10.1994, 59.6703],
-  [10.2019, 59.6693],
-  [10.2021, 59.6692],
-  [10.2028, 59.669],
-  [10.203, 59.6689],
-  [10.2037, 59.6688],
-  [10.205, 59.6686],
-  [10.2059, 59.6685],
-  [10.2069, 59.6684],
-  [10.2092, 59.6682],
-  [10.2098, 59.6681],
-  [10.2101, 59.668],
-  [10.2102, 59.668],
-  [10.2101, 59.6676],
-  [10.21, 59.6672],
-  [10.2101, 59.667],
-  [10.2102, 59.6668],
-  [10.2104, 59.6666],
-  [10.2106, 59.6665],
-  [10.211, 59.6665],
-  [10.2114, 59.6666],
-  [10.2117, 59.6667],
-  [10.212, 59.6669],
-  [10.2123, 59.667],
-  [10.2127, 59.6671],
-  [10.2129, 59.6671],
-  [10.2131, 59.6671],
-  [10.2133, 59.667],
-  [10.2134, 59.6669],
-  [10.2135, 59.6668],
-  [10.2137, 59.6663],
-  [10.2137, 59.6663],
-  [10.2141, 59.666],
-  [10.2142, 59.6659],
-  [10.2143, 59.6658],
-  [10.2144, 59.6658],
-  [10.2146, 59.6657],
-  [10.2147, 59.6657],
-  [10.2149, 59.6656],
-  [10.215, 59.6654],
-  [10.2152, 59.6652],
-  [10.2152, 59.6647],
-  [10.2153, 59.6644],
-  [10.2156, 59.664],
-  [10.2157, 59.664],
-  [10.218, 59.6639],
-  [10.2188, 59.6638],
-  [10.219, 59.6638],
-  [10.219, 59.6639],
-  [10.219, 59.664],
-  [10.2191, 59.6643],
-  [10.2197, 59.6644],
-  [10.2203, 59.6644],
-  [10.2203, 59.6644],
-  [10.2204, 59.6647],
-  [10.2204, 59.665],
-  [10.2204, 59.6652],
-  [10.2205, 59.6651],
-  [10.2209, 59.6652],
-  [10.2222, 59.6655],
-  [10.2232, 59.6653],
-  [10.2232, 59.6654],
-  [10.2234, 59.6655],
-  [10.2234, 59.6656],
-  [10.2233, 59.6657],
-  [10.2234, 59.666],
-  [10.2236, 59.6662],
-  [10.2237, 59.6663],
-  [10.2236, 59.6665],
-  [10.2236, 59.6665],
-  [10.2235, 59.6666],
-  [10.2234, 59.6668],
-  [10.2233, 59.667],
-  [10.2234, 59.6671],
-  [10.2233, 59.6673],
-  [10.2231, 59.6675],
-  [10.2228, 59.6676],
-  [10.2227, 59.6679],
-  [10.2226, 59.6681],
-  [10.2226, 59.6681],
-  [10.2226, 59.6683],
-  [10.2226, 59.6684],
-  [10.2227, 59.6684],
-  [10.2229, 59.6686],
-  [10.223, 59.6687],
-  [10.223, 59.6688],
-  [10.223, 59.6688],
-  [10.223, 59.6689],
-  [10.223, 59.6689],
-  [10.2231, 59.669],
-  [10.223, 59.6691],
-  [10.2231, 59.6692],
-  [10.223, 59.6693],
-  [10.2232, 59.6694],
-  [10.2236, 59.6698],
-  [10.2238, 59.6702],
-  [10.2238, 59.6702],
-  [10.224, 59.6704],
-  [10.2241, 59.6705],
-  [10.2241, 59.6706],
-  [10.2242, 59.6707],
-  [10.2241, 59.6709],
-  [10.224, 59.671],
-  [10.2245, 59.6713],
-  [10.2251, 59.6714],
-  [10.2252, 59.6714],
-  [10.2252, 59.6714],
-  [10.2257, 59.6712],
-  [10.2272, 59.6709],
-  [10.2281, 59.671],
-  [10.2293, 59.671],
-  [10.2306, 59.671],
-  [10.2307, 59.6707],
-  [10.231, 59.6701],
-  [10.231, 59.6699],
-  [10.2323, 59.6697],
-  [10.2332, 59.6695],
-  [10.2353, 59.6685],
-  [10.2374, 59.6671],
-  [10.2374, 59.6671],
-  [10.2383, 59.6666],
-  [10.2386, 59.6664],
-  [10.2388, 59.6667],
-  [10.2396, 59.6678],
-  [10.2397, 59.6678],
-  [10.24, 59.668],
-  [10.2405, 59.6684],
-  [10.2406, 59.6684],
-  [10.2407, 59.6685],
-  [10.2408, 59.6686],
-  [10.2408, 59.6687],
-  [10.2408, 59.6689],
-  [10.2408, 59.6694],
-  [10.2409, 59.6694],
-  [10.2411, 59.6695],
-  [10.2414, 59.6698],
-  [10.2415, 59.6699],
-  [10.2424, 59.6705],
-  [10.2424, 59.6707],
-  [10.2424, 59.6709],
-  [10.2424, 59.6711],
-  [10.2424, 59.6713],
-  [10.2427, 59.6715],
-  [10.243, 59.6719],
-  [10.2431, 59.6721],
-  [10.2435, 59.6725],
-  [10.2439, 59.6726],
-  [10.244, 59.6727],
-  [10.2444, 59.6728],
-  [10.2448, 59.6729],
-  [10.2452, 59.673],
-  [10.2454, 59.6731],
-  [10.2459, 59.6732],
-  [10.246, 59.6732],
-  [10.2462, 59.6732],
-  [10.2464, 59.6732],
-  [10.2466, 59.6732],
-  [10.2469, 59.6733],
-  [10.2469, 59.6733],
-  [10.2472, 59.6733],
-  [10.2475, 59.6732],
-  [10.2476, 59.6732],
-  [10.248, 59.6733],
-  [10.2483, 59.6733],
-  [10.2492, 59.6734],
-  [10.2496, 59.6736],
-  [10.2499, 59.6735],
-  [10.2502, 59.6736],
-  [10.2503, 59.6737],
-  [10.2502, 59.6737],
-  [10.25, 59.6738],
-  [10.25, 59.6738],
-  [10.2501, 59.6739],
-  [10.2503, 59.6739],
-  [10.2507, 59.6738],
-  [10.251, 59.674],
-  [10.2512, 59.674],
-  [10.2514, 59.6739],
-  [10.2516, 59.6737],
-  [10.2519, 59.6738],
-  [10.2519, 59.6737],
-  [10.2521, 59.6737],
-  [10.2525, 59.6736],
-  [10.2525, 59.6735],
-  [10.2526, 59.6735],
-  [10.2527, 59.6735],
-  [10.2529, 59.6735],
-  [10.253, 59.6736],
-  [10.2533, 59.6736],
-  [10.2534, 59.6735],
-  [10.2537, 59.6734],
-  [10.2538, 59.6734],
-  [10.2538, 59.6733],
-  [10.2539, 59.6733],
-  [10.2542, 59.6731],
-  [10.2543, 59.6731],
-  [10.2544, 59.6732],
-  [10.2546, 59.6734],
-  [10.2548, 59.6735],
-  [10.2551, 59.6736],
-  [10.2556, 59.6734],
-  [10.256, 59.6735],
-  [10.2562, 59.6735],
-  [10.2563, 59.6736],
-  [10.2562, 59.6737],
-  [10.2565, 59.6738],
-  [10.2568, 59.6739],
-  [10.2571, 59.6738],
-  [10.2576, 59.6737],
-  [10.258, 59.6736],
-  [10.2584, 59.6735],
-  [10.2584, 59.6735],
-  [10.2589, 59.6735],
-  [10.2596, 59.6738],
-  [10.2599, 59.6738],
-  [10.2607, 59.6738],
-  [10.2614, 59.6739],
-  [10.2616, 59.6739],
-  [10.2617, 59.6738],
-  [10.2623, 59.6738],
-  [10.2628, 59.6737],
-  [10.2631, 59.6737],
-  [10.2635, 59.6736],
-  [10.2641, 59.6735],
-  [10.2644, 59.6732],
-  [10.2648, 59.673],
-  [10.2651, 59.673],
-  [10.2663, 59.6729],
-  [10.2671, 59.6728],
-  [10.27, 59.6723],
-  [10.2714, 59.6711],
-  [10.272, 59.6709],
-  [10.2741, 59.671],
-  [10.2751, 59.6712],
-  [10.2768, 59.6712],
-  [10.2778, 59.6711],
-  [10.279, 59.6711],
-  [10.2796, 59.6715],
-  [10.2801, 59.6717],
-  [10.2814, 59.6716],
-  [10.2817, 59.6716],
-  [10.2829, 59.6716],
-  [10.2837, 59.6717],
-  [10.2846, 59.6714],
-  [10.2854, 59.671],
-  [10.2864, 59.6705],
-  [10.287, 59.6702],
-  [10.2871, 59.6702],
-  [10.2887, 59.6699],
-  [10.2887, 59.6699],
-  [10.289, 59.6698],
-  [10.2901, 59.6697],
-  [10.2913, 59.6697],
-  [10.294, 59.6697],
-  [10.2953, 59.6697],
-  [10.2956, 59.6696],
-  [10.2978, 59.6694],
-  [10.2996, 59.6691],
-  [10.3, 59.669],
-  [10.3011, 59.6687],
-  [10.3029, 59.6682],
-  [10.3053, 59.6701],
-  [10.3069, 59.6715],
-  [10.308, 59.6723],
-  [10.3083, 59.6728],
-  [10.3087, 59.6733],
-  [10.3075, 59.6749],
-  [10.3085, 59.6763],
-  [10.3103, 59.6776],
-  [10.3103, 59.6786],
-  [10.3103, 59.6796],
-  [10.3099, 59.6797],
-  [10.3099, 59.6798],
-  [10.3098, 59.6799],
-  [10.3099, 59.6802],
-  [10.3118, 59.681],
-  [10.3116, 59.6815],
-  [10.3116, 59.6817],
-  [10.3117, 59.6818],
-  [10.3116, 59.682],
-  [10.3117, 59.6822],
-  [10.3118, 59.6823],
-  [10.3118, 59.6824],
-  [10.3118, 59.6825],
-  [10.3122, 59.6826],
-  [10.3123, 59.6826],
-  [10.3124, 59.6826],
-  [10.3134, 59.6827],
-  [10.3136, 59.6832],
-  [10.3139, 59.6832],
-  [10.3139, 59.6834],
-  [10.3143, 59.6839],
-  [10.3147, 59.6842],
-  [10.3145, 59.6845],
-  [10.3149, 59.6847],
-  [10.3154, 59.6849],
-  [10.3155, 59.685],
-  [10.3157, 59.6851],
-  [10.3159, 59.6852],
-  [10.3161, 59.6853],
-  [10.3161, 59.6854],
-  [10.3166, 59.6856],
-  [10.3168, 59.6857],
-  [10.317, 59.6858],
-  [10.3172, 59.686],
-  [10.3173, 59.686],
-  [10.3174, 59.6862],
-  [10.3177, 59.6863],
-  [10.3179, 59.6863],
-  [10.3182, 59.6865],
-  [10.3185, 59.6866],
-  [10.3187, 59.6867],
-  [10.319, 59.6867],
-  [10.3192, 59.6869],
-  [10.3194, 59.687],
-  [10.3195, 59.687],
-  [10.3199, 59.6872],
-  [10.3203, 59.6873],
-  [10.3206, 59.6875],
-  [10.3209, 59.6876],
-  [10.3211, 59.6878],
-  [10.3212, 59.6878],
-  [10.3213, 59.6879],
-  [10.3214, 59.6879],
-  [10.3215, 59.688],
-  [10.3215, 59.688],
-  [10.3215, 59.6881],
-  [10.3213, 59.6882],
-  [10.3213, 59.6882],
-  [10.3213, 59.6883],
-  [10.3213, 59.6884],
-  [10.3214, 59.6885],
-  [10.3213, 59.6886],
-  [10.3218, 59.6889],
-  [10.322, 59.689],
-  [10.322, 59.6892],
-  [10.3222, 59.6893],
-  [10.3225, 59.6894],
-  [10.3228, 59.6895],
-  [10.3228, 59.6896],
-  [10.3226, 59.6898],
-  [10.3222, 59.6901],
-  [10.3218, 59.6903],
-  [10.3217, 59.6904],
-  [10.3213, 59.6904],
-  [10.3208, 59.6905],
-  [10.3208, 59.6906],
-  [10.3207, 59.6908],
-  [10.3205, 59.6909],
-  [10.3203, 59.6909],
-  [10.3201, 59.6911],
-  [10.3199, 59.6911],
-  [10.3199, 59.6912],
-  [10.3199, 59.6913],
-  [10.3202, 59.6923],
-  [10.3234, 59.7019],
-  [10.3635, 59.6977],
-  [10.3917, 59.6793],
-  [10.4031, 59.6717],
-  [10.4104, 59.6671],
-  [10.4101, 59.6455],
-  [10.4245, 59.6253],
-  [10.412, 59.618],
-  [10.4108, 59.6092],
-  [10.4233, 59.6004],
-  [10.4235, 59.5982],
-  [10.4116, 59.5851],
-  [10.4178, 59.5672],
-  [10.398, 59.5304],
-  [10.3937, 59.5147],
-  [10.4558, 59.4799],
-  [10.5431, 59.4573],
-  [10.5367, 59.4328],
-  [10.543, 59.4153],
-  [10.544, 59.4123],
-  [10.5618, 59.4038],
-  [10.5842, 59.3931],
-  [10.5953, 59.3741],
-  [10.5952, 59.3715],
-  [10.5945, 59.3582],
-  [10.5944, 59.3562],
-  [10.5916, 59.3535],
-  [10.5782, 59.3409],
-  [10.5895, 59.3017],
-  [10.5908, 59.2972],
-  [10.5903, 59.2685],
-  [10.5972, 59.261],
-  [10.605, 59.2523],
-  [10.6124, 59.2423],
-  [10.6164, 59.237],
-  [10.6416, 59.2029],
-  [10.655, 59.192],
-  [10.6536, 59.1707],
-  [10.6514, 59.1386],
-  [10.651, 59.1328],
-  [10.6376, 59.1228],
-  [10.6227, 59.1117],
-  [10.6509, 59.0759],
-  [10.675, 59.0453],
-  [10.6604, 59.0248],
-  [10.6514, 59.0049],
-  [10.6285, 58.9546],
-  [10.5931, 58.761],
-  [10.5762, 58.7618],
-  [10.3638, 58.7711],
-  [10.3594, 58.7713],
-  [10.2855, 58.7744],
-  [10.2709, 58.775],
-  [10.2634, 58.7739],
-  [10.2526, 58.7723],
-  [10.1495, 58.7583],
-  [10.0738, 58.7479],
-  [10.0593, 58.7441],
-  [10.0047, 58.7297],
-  [9.988, 58.7253],
-  [9.9698, 58.7205],
-  [9.8611, 58.8314],
-  [9.8537, 58.8389],
-  [9.8536, 58.839],
-  [9.7952, 58.8982],
-  [9.7831, 58.9358],
-  [9.7692, 58.9844],
-  [9.7676, 58.99],
-  [9.765, 58.9992],
-  [9.7645, 59.0011],
-  [9.7648, 59.01],
-  [9.7811, 59.0118],
-  [9.7825, 59.03],
-  [9.8346, 58.9978],
-  [9.8355, 59.0129],
-  [9.8362, 59.0231],
-  [9.8231, 59.0316],
-  [9.8224, 59.0322],
-  [9.8217, 59.0323],
-  [9.8215, 59.0323],
-  [9.8211, 59.0324],
-  [9.8207, 59.033],
-  [9.8208, 59.0331],
-  [9.8128, 59.034],
-  [9.8129, 59.036],
-  [9.8133, 59.0363],
-  [9.8134, 59.0363],
-  [9.8135, 59.0364],
-  [9.8135, 59.0364],
-  [9.8146, 59.0371],
-  [9.8146, 59.0372],
-  [9.8148, 59.0378],
-  [9.815, 59.0382],
-  [9.8155, 59.0388],
-  [9.8158, 59.0391],
-  [9.8163, 59.0396],
-  [9.8165, 59.0398],
-  [9.817, 59.0403],
-  [9.8172, 59.0405],
-  [9.8175, 59.0408],
-  [9.8178, 59.0408],
-  [9.818, 59.0408],
-  [9.8181, 59.0409],
-  [9.8186, 59.041],
-  [9.8193, 59.0411],
-  [9.8203, 59.0414],
-  [9.8216, 59.0417],
-  [9.8219, 59.0418],
-  [9.8227, 59.042],
-  [9.8239, 59.0423],
-  [9.8245, 59.0425],
-  [9.8249, 59.0425],
-  [9.8258, 59.0423],
-  [9.8262, 59.0419],
-  [9.8267, 59.0414],
-  [9.8273, 59.0412],
-  [9.8279, 59.0408],
-  [9.8285, 59.0404],
-  [9.8292, 59.0401],
-  [9.8298, 59.0398],
-  [9.8304, 59.0395],
-  [9.8321, 59.0387],
-  [9.8325, 59.0384],
-  [9.833, 59.0382],
-  [9.8333, 59.038],
-  [9.8333, 59.038],
-  [9.8333, 59.038],
-  [9.8335, 59.0379],
-  [9.8336, 59.0378],
-  [9.8341, 59.0376],
-  [9.8347, 59.0372],
-  [9.8392, 59.0441],
-  [9.8392, 59.0442],
-  [9.8392, 59.0444],
-  [9.8392, 59.0445],
-  [9.8392, 59.0445],
-  [9.8392, 59.0446],
-  [9.8392, 59.0448],
-  [9.8392, 59.0449],
-  [9.8392, 59.0449],
-  [9.8392, 59.045],
-  [9.8391, 59.045],
-  [9.8391, 59.0451],
-  [9.8391, 59.0451],
-  [9.8391, 59.0452],
-  [9.8391, 59.0453],
-  [9.8391, 59.0453],
-  [9.8391, 59.0454],
-  [9.8391, 59.0454],
-  [9.8391, 59.0454],
-  [9.8392, 59.0455],
-  [9.8392, 59.0455],
-  [9.8392, 59.0456],
-  [9.8392, 59.0456],
-  [9.8392, 59.0457],
-  [9.8392, 59.0457],
-  [9.8392, 59.0458],
-  [9.8393, 59.0458],
-  [9.8394, 59.0459],
-  [9.8393, 59.0459],
-  [9.8394, 59.0459],
-  [9.8394, 59.046],
-  [9.8394, 59.0461],
-  [9.8394, 59.0462],
-  [9.8394, 59.0463],
-  [9.8395, 59.0463],
-  [9.8395, 59.0464],
-  [9.8396, 59.0466],
-  [9.8396, 59.0467],
-  [9.8397, 59.0469],
-  [9.8398, 59.0471],
-  [9.8398, 59.0472],
-  [9.8399, 59.0477],
-  [9.84, 59.0485],
-  [9.84, 59.0485],
-  [9.8399, 59.0485],
-  [9.8399, 59.0487],
-  [9.8399, 59.0487],
-  [9.8399, 59.0488],
-  [9.8399, 59.0488],
-  [9.8398, 59.0489],
-  [9.8398, 59.049],
-  [9.8398, 59.049],
-  [9.8398, 59.0491],
-  [9.8398, 59.0492],
-  [9.8397, 59.0492],
-  [9.8397, 59.0493],
-  [9.8398, 59.0493],
-  [9.8398, 59.0494],
-  [9.8398, 59.0494],
-  [9.8398, 59.0495],
-  [9.8398, 59.0495],
-  [9.8398, 59.0496],
-  [9.8398, 59.0497],
-  [9.8398, 59.0497],
-  [9.8398, 59.0497],
-  [9.8398, 59.0498],
-  [9.8398, 59.0498],
-  [9.8398, 59.0499],
-  [9.8398, 59.05],
-  [9.8397, 59.05],
-  [9.8397, 59.05],
-  [9.8396, 59.0501],
-  [9.8396, 59.0501],
-  [9.8395, 59.0502],
-  [9.8394, 59.0503],
-  [9.8393, 59.0504],
-  [9.8393, 59.0504],
-  [9.8393, 59.0505],
-  [9.8393, 59.0505],
-  [9.8393, 59.0505],
-  [9.8392, 59.0506],
-  [9.8392, 59.0506],
-  [9.8392, 59.0507],
-  [9.8392, 59.0507],
-  [9.8392, 59.0507],
-  [9.8391, 59.0508],
-  [9.8391, 59.0508],
-  [9.8391, 59.0509],
-  [9.8391, 59.0509],
-  [9.839, 59.0509],
-  [9.839, 59.051],
-  [9.839, 59.051],
-  [9.839, 59.0511],
-  [9.8389, 59.0511],
-  [9.8389, 59.0511],
-  [9.8389, 59.0512],
-  [9.839, 59.0512],
-  [9.839, 59.0512],
-  [9.8389, 59.0512],
-  [9.8389, 59.0513],
-  [9.8389, 59.0513],
-  [9.8388, 59.0513],
-  [9.8388, 59.0513],
-  [9.8387, 59.0514],
-  [9.8387, 59.0514],
-  [9.8386, 59.0514],
-  [9.8387, 59.0515],
-  [9.8387, 59.0515],
-  [9.8387, 59.0515],
-  [9.8388, 59.0516],
-  [9.8388, 59.0516],
-  [9.8388, 59.0516],
-  [9.8388, 59.0517],
-  [9.8388, 59.0517],
-  [9.8388, 59.0517],
-  [9.8388, 59.0518],
-  [9.8388, 59.0518],
-  [9.8387, 59.0518],
-  [9.8387, 59.0519],
-  [9.8387, 59.0519],
-  [9.8387, 59.0519],
-  [9.8388, 59.0519],
-  [9.8388, 59.0519],
-  [9.8389, 59.052],
-  [9.8389, 59.052],
-  [9.839, 59.052],
-  [9.839, 59.052],
-  [9.8391, 59.0521],
-  [9.8391, 59.0521],
-  [9.8391, 59.0522],
-  [9.8392, 59.0522],
-  [9.8392, 59.0523],
-  [9.8392, 59.0523],
-  [9.8392, 59.0524],
-  [9.8392, 59.0524],
-  [9.8392, 59.0525],
-  [9.8392, 59.0526],
-  [9.8392, 59.0526],
-  [9.8392, 59.0526],
-  [9.8392, 59.0526],
-  [9.8392, 59.0526],
-  [9.8392, 59.0527],
-  [9.8392, 59.0527],
-  [9.8392, 59.0527],
-  [9.8391, 59.0528],
-  [9.8391, 59.0528],
-  [9.8392, 59.0528],
-  [9.8392, 59.0528],
-  [9.8392, 59.0528],
-  [9.8392, 59.0529],
-  [9.8392, 59.0529],
-  [9.8392, 59.053],
-  [9.8391, 59.053],
-  [9.8391, 59.0531],
-  [9.8391, 59.0531],
-  [9.8392, 59.0532],
-  [9.8392, 59.0532],
-  [9.8392, 59.0534],
-  [9.8392, 59.0535],
-  [9.8392, 59.0535],
-  [9.8392, 59.0536],
-  [9.8392, 59.0536],
-  [9.8392, 59.0536],
-  [9.8392, 59.0537],
-  [9.8392, 59.0537],
-  [9.8391, 59.0538],
-  [9.8391, 59.0539],
-  [9.8391, 59.0541],
-  [9.8391, 59.0542],
-  [9.839, 59.0545],
-  [9.8389, 59.0554],
-  [9.8389, 59.0556],
-  [9.8389, 59.0559],
-  [9.8391, 59.0563],
-  [9.8393, 59.0567],
-  [9.8398, 59.057],
-  [9.8401, 59.0574],
-  [9.8404, 59.0582],
-  [9.8401, 59.0587],
-  [9.84, 59.0587],
-  [9.8401, 59.0589],
-  [9.8403, 59.0593],
-  [9.8406, 59.0598],
-  [9.841, 59.0607],
-  [9.8417, 59.062],
-  [9.8419, 59.0624],
-  [9.842, 59.0628],
-  [9.8419, 59.0634],
-  [9.8418, 59.0638],
-  [9.8417, 59.0642],
-  [9.8414, 59.065],
-  [9.8413, 59.0654],
-  [9.8412, 59.0656],
-  [9.841, 59.0661],
-  [9.8408, 59.0664],
-  [9.8406, 59.0669],
-  [9.8406, 59.067],
-  [9.8406, 59.0675],
-  [9.8406, 59.0679],
-  [9.8405, 59.0684],
-  [9.8405, 59.0688],
-  [9.8404, 59.0692],
-  [9.84, 59.0696],
-  [9.8399, 59.07],
-  [9.8397, 59.0703],
-  [9.8394, 59.0707],
-  [9.8391, 59.0711],
-  [9.839, 59.0712],
-  [9.8389, 59.0716],
-  [9.8386, 59.072],
-  [9.8383, 59.0725],
-  [9.8389, 59.0725],
-  [9.8398, 59.0725],
-  [9.8406, 59.0725],
-  [9.8416, 59.0725],
-  [9.8422, 59.0726],
-  [9.8424, 59.0726],
-  [9.8433, 59.0726],
-  [9.8442, 59.0726],
-  [9.845, 59.0726],
-  [9.8453, 59.0726],
-  [9.846, 59.0727],
-  [9.8468, 59.0728],
-  [9.8477, 59.0729],
-  [9.8484, 59.0729],
-  [9.8486, 59.0731],
-  [9.8492, 59.0734],
-  [9.8498, 59.0738],
-  [9.8499, 59.0739],
-  [9.8501, 59.0741],
-  [9.8502, 59.0744],
-  [9.8503, 59.0746],
-  [9.8504, 59.075],
-  [9.8505, 59.0754],
-  [9.8509, 59.0762],
-  [9.851, 59.0766],
-  [9.8513, 59.0773],
-  [9.8513, 59.0774],
-  [9.8513, 59.0775],
-  [9.8513, 59.0775],
-  [9.8514, 59.0776],
-  [9.8514, 59.0776],
-  [9.8514, 59.0777],
-  [9.8515, 59.0778],
-  [9.8515, 59.0779],
-  [9.8515, 59.078],
-  [9.8515, 59.078],
-  [9.8522, 59.0805],
-  [9.8523, 59.0808],
-  [9.8524, 59.081],
-  [9.8524, 59.0813],
-  [9.8527, 59.0819],
-  [9.8528, 59.0822],
-  [9.8529, 59.0826],
-  [9.8531, 59.0834],
-  [9.8533, 59.0842],
-  [9.8534, 59.0844],
-  [9.8534, 59.0848],
-  [9.8535, 59.0852],
-  [9.8538, 59.0873],
-  [9.854, 59.0876],
-  [9.8541, 59.0879],
-  [9.854, 59.0883],
-  [9.8536, 59.0894],
-  [9.8535, 59.0896],
-  [9.8535, 59.0899],
-  [9.8535, 59.0904],
-  [9.8538, 59.091],
-  [9.8538, 59.0915],
-  [9.8541, 59.092],
-  [9.8542, 59.0925],
-  [9.8544, 59.0928],
-  [9.8545, 59.0931],
-  [9.8546, 59.0933],
-  [9.8548, 59.0936],
-  [9.855, 59.0938],
-  [9.855, 59.0941],
-  [9.855, 59.0944],
-  [9.8555, 59.0948],
-  [9.8555, 59.0953],
-  [9.8555, 59.0957],
-  [9.8556, 59.0961],
-  [9.8558, 59.0965],
-  [9.8558, 59.0966],
-  [9.8559, 59.0968],
-  [9.8559, 59.097],
-  [9.856, 59.0974],
-  [9.856, 59.0976],
-  [9.8561, 59.0979],
-  [9.8561, 59.0984],
-  [9.8561, 59.0988],
-  [9.8561, 59.0992],
-  [9.8557, 59.0997],
-  [9.8551, 59.1004],
-  [9.8545, 59.101],
-  [9.8542, 59.1014],
-  [9.854, 59.1017],
-  [9.8538, 59.1019],
-  [9.8537, 59.1021],
-  [9.8535, 59.1025],
-  [9.8533, 59.1029],
-  [9.8531, 59.1033],
-  [9.8527, 59.1041],
-  [9.8523, 59.1049],
-  [9.8522, 59.1053],
-  [9.8519, 59.1056],
-  [9.8517, 59.106],
-  [9.8515, 59.1064],
-  [9.8512, 59.1067],
-  [9.8508, 59.1076],
-  [9.8505, 59.1079],
-  [9.8498, 59.1087],
-  [9.8497, 59.109],
-  [9.8496, 59.1091],
-  [9.8495, 59.1095],
-  [9.8495, 59.1096],
-  [9.8494, 59.1098],
-  [9.8493, 59.11],
-  [9.8491, 59.1105],
-  [9.8489, 59.1109],
-  [9.8487, 59.1113],
-  [9.8486, 59.1116],
-  [9.8485, 59.1117],
-  [9.8479, 59.1122],
-  [9.8476, 59.1125],
-  [9.8469, 59.1132],
-  [9.8462, 59.1137],
-  [9.846, 59.1139],
-  [9.8454, 59.1143],
-  [9.8444, 59.115],
-  [9.8441, 59.1152],
-  [9.8431, 59.1158],
-  [9.8425, 59.1162],
-  [9.8422, 59.1164],
-  [9.8419, 59.1166],
-  [9.8415, 59.1169],
-  [9.8405, 59.1176],
-  [9.8398, 59.118],
-  [9.8394, 59.1183],
-  [9.8391, 59.1184],
-  [9.8386, 59.1189],
-  [9.8381, 59.1192],
-  [9.8375, 59.1196],
-  [9.8369, 59.12],
-  [9.8364, 59.1203],
-  [9.8359, 59.1207],
-  [9.8354, 59.121],
-  [9.8349, 59.1215],
-  [9.8343, 59.1218],
-  [9.8338, 59.1222],
-  [9.8332, 59.1225],
-  [9.8326, 59.1229],
-  [9.8321, 59.1233],
-  [9.8295, 59.1251],
-  [9.8288, 59.1254],
-  [9.8282, 59.1257],
-  [9.8262, 59.1267],
-  [9.8243, 59.1276],
-  [9.8236, 59.128],
-  [9.8225, 59.1285],
-  [9.8219, 59.1288],
-  [9.8216, 59.129],
-  [9.8206, 59.1295],
-  [9.8212, 59.1303],
-  [9.8223, 59.1315],
-  [9.8223, 59.1315],
-  [9.8224, 59.1319],
-  [9.8226, 59.1321],
-  [9.8227, 59.1323],
-  [9.8232, 59.1327],
-  [9.8236, 59.1331],
-  [9.8239, 59.1334],
-  [9.825, 59.1346],
-  [9.8253, 59.1349],
-  [9.8259, 59.1355],
-  [9.8265, 59.1361],
-  [9.8268, 59.1364],
-  [9.8277, 59.1375],
-  [9.8282, 59.138],
-  [9.8288, 59.1388],
-  [9.8292, 59.1395],
-  [9.8293, 59.1396],
-  [9.8293, 59.14],
-  [9.8294, 59.1412],
-  [9.8293, 59.1414],
-  [9.8293, 59.1419],
-  [9.8292, 59.1428],
-  [9.8292, 59.1429],
-  [9.8291, 59.1437],
-  [9.8289, 59.1456],
-  [9.8288, 59.1462],
-  [9.8287, 59.1473],
-  [9.8287, 59.1478],
-  [9.8286, 59.1483],
-  [9.8285, 59.1491],
-  [9.8284, 59.1499],
-  [9.828, 59.1524],
-  [9.8278, 59.1537],
-  [9.8277, 59.1538],
-  [9.8277, 59.154],
-  [9.8276, 59.1543],
-  [9.8276, 59.1545],
-  [9.8274, 59.1555],
-  [9.8272, 59.156],
-  [9.8269, 59.1568],
-  [9.8267, 59.1569],
-  [9.8265, 59.157],
-  [9.8255, 59.1574],
-  [9.8248, 59.1576],
-  [9.8259, 59.16],
-  [9.8279, 59.1614],
-  [9.8293, 59.1624],
-  [9.8313, 59.1639],
-  [9.8346, 59.164],
-  [9.841, 59.164],
-  [9.8419, 59.164],
-  [9.8428, 59.164],
-  [9.8437, 59.164],
-  [9.8499, 59.1642],
-  [9.8508, 59.1648],
-  [9.8509, 59.1648],
-  [9.851, 59.1649],
-  [9.851, 59.1649],
-  [9.851, 59.1649],
-  [9.851, 59.1649],
-  [9.8529, 59.1647],
-  [9.8542, 59.1643],
-  [9.8547, 59.1643],
-  [9.8682, 59.1671],
-  [9.8855, 59.1693],
-  [9.8964, 59.177],
-  [9.8962, 59.1779],
-  [9.8924, 59.1834],
-  [9.8896, 59.1878],
-  [9.885, 59.1931],
-  [9.8786, 59.1978],
-  [9.875, 59.2004],
-  [9.8718, 59.2028],
-  [9.8716, 59.2029],
-  [9.8711, 59.2033],
-  [9.8689, 59.2049],
-  [9.8682, 59.2057],
-  [9.8672, 59.2069],
-  [9.8666, 59.2078],
-  [9.8655, 59.2093],
-  [9.8653, 59.2095],
-  [9.8649, 59.2101],
-  [9.8635, 59.2115],
-  [9.8628, 59.2121],
-  [9.8617, 59.2127],
-  [9.86, 59.2137],
-  [9.8593, 59.214],
-  [9.859, 59.2141],
-  [9.8579, 59.215],
-  [9.8574, 59.2154],
-  [9.8572, 59.2155],
-  [9.8574, 59.2157],
-  [9.8578, 59.2159],
-  [9.8578, 59.2159],
-  [9.858, 59.216],
-  [9.8604, 59.2159],
-  [9.8624, 59.2164],
-  [9.8649, 59.2171],
-  [9.8664, 59.2176],
-  [9.8684, 59.2182],
-  [9.8704, 59.2187],
-  [9.8713, 59.2189],
-  [9.8779, 59.2207],
-  [9.8781, 59.2207],
-  [9.8775, 59.2214],
-  [9.877, 59.222],
-  [9.8767, 59.2225],
-  [9.8762, 59.2232],
-  [9.876, 59.2237],
-  [9.8762, 59.2247],
-  [9.8764, 59.2253],
-  [9.8763, 59.2257],
-  [9.8758, 59.2266],
-  [9.8755, 59.227],
-  [9.8752, 59.2277],
-  [9.8747, 59.2281],
-  [9.8747, 59.2287],
-  [9.8746, 59.2292],
-  [9.8742, 59.2335],
-  [9.8734, 59.2347],
-  [9.8733, 59.235],
-  [9.8731, 59.2355],
-  [9.8728, 59.236],
-  [9.8726, 59.2366],
-  [9.8723, 59.2374],
-  [9.8718, 59.2384],
-  [9.8716, 59.2389],
-  [9.8714, 59.2394],
-  [9.8714, 59.2403],
-  [9.8715, 59.2408],
-  [9.8715, 59.2419],
-  [9.8716, 59.2433],
-  [9.8716, 59.2438],
-  [9.8717, 59.2449],
-  [9.8717, 59.2454],
-  [9.8717, 59.2457],
-  [9.8717, 59.2459],
-  [9.8718, 59.2462],
-  [9.8717, 59.2463],
-  [9.8712, 59.2477],
-  [9.8711, 59.2479],
-  [9.871, 59.2481],
-  [9.8709, 59.2485],
-  [9.8704, 59.2497],
-  [9.8695, 59.2519],
-  [9.8668, 59.253],
-  [9.864, 59.2543],
-  [9.8626, 59.2549],
-  [9.8618, 59.2553],
-  [9.8611, 59.2556],
-  [9.8605, 59.2559],
-  [9.8557, 59.2591],
-  [9.8551, 59.2594],
-  [9.8547, 59.2595],
-  [9.8533, 59.2602],
-  [9.8529, 59.2604],
-  [9.8521, 59.2609],
-  [9.8518, 59.261],
-  [9.8502, 59.2619],
-  [9.8495, 59.2623],
-  [9.8473, 59.2646],
-  [9.8468, 59.2649],
-];
