+++ src/test/mock/unequippedStopPlace.js
